+++ pyre/pyre-jit-trace/src/jitcode_dispatch/mod.rs
@@ -175,6 +175,12 @@ use majit_metainterp::{TraceCtx, default_effect_info};
 // jitcode_dispatch submodules (extracted from this file). Their `pub`
 // items are re-exported so `crate::jitcode_dispatch::` paths stay stable.
 // __SUBMODULES__
+mod vable_ops;
+pub use vable_ops::*;
+mod heapcache_ops;
+pub use heapcache_ops::*;
+mod bridge_subwalk;
+pub use bridge_subwalk::*;
 mod arith;
 pub use arith::*;
 mod diag;
@@ -2578,1666 +2584,6 @@ fn seed_standing_exception_for_walk(sym: &mut crate::state::PyreSym, trace_ctx: 
     }
 }
 
-#[allow(clippy::too_many_arguments)]
-pub fn dispatch_via_miframe(
-    miframe: &mut MIFrame,
-    session: &std::cell::RefCell<WalkSession>,
-    jitcode_code: &[u8],
-    position: usize,
-    descr_refs: &[DescrRef],
-    raw_descrs: RawDescrPool,
-    is_authoritative_executor: bool,
-    sub_jitcode_lookup: &SubJitCodeLookup,
-    done_with_this_frame_descr_ref: DescrRef,
-    done_with_this_frame_descr_int: DescrRef,
-    done_with_this_frame_descr_float: DescrRef,
-    done_with_this_frame_descr_void: DescrRef,
-    exit_frame_with_exception_descr_ref: DescrRef,
-    is_top_level: bool,
-    // PyPy `pyjitpl.py:171-176 MIFrame.__init__` analog: the
-    // top-level jitcode's per-bank register count.  `dispatch_via_miframe`
-    // allocates fresh `Vec<OpRef>`s sized to `top_num_regs_* +
-    // top_constants_*.len()` — replacing the prior TODO that
-    // reused `sym.registers_r` (a Python locals/stack mirror) as the
-    // MIFrame register file.  The codewriter-compiled arm jitcode
-    // expects `R[0]_r = handler = MIFrame self ptr`, which the
-    // `argboxes_*` parameters supply via the `setup_call` analog
-    // below.
-    top_num_regs_r: usize,
-    top_num_regs_i: usize,
-    top_num_regs_f: usize,
-    // Top-level jitcode's per-bank constant pool — seeded into
-    // register slots `[num_regs_*, num_regs_* + constants_*.len())`
-    // per `pyjitpl.py:98-119 copy_constants`.
-    top_constants_r: &[i64],
-    top_constants_i: &[i64],
-    top_constants_f: &[i64],
-    // PyPy `pyjitpl.py:188-200 setup_call(argboxes)` analog.
-    // `argboxes_*[i]` is written to `registers_*[i]` before walking.
-    // Production callers supply `argboxes_r = [const_ref(miframe_ptr)]`
-    // so the codewriter-compiled arm finds the MIFrame self ptr at
-    // `R[0]_r`.
-    argboxes_r: &[OpRef],
-    argboxes_i: &[OpRef],
-    argboxes_f: &[OpRef],
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Extract raw pointers before any borrow. `miframe.ctx` and
-    // `miframe.sym` are `*mut`, distinct objects (the trace
-    // recorder vs. the symbolic frame state) — distinct pointers
-    // means dereferencing both simultaneously is sound.
-    let ctx_ptr = miframe.ctx;
-    let sym_ptr = miframe.sym;
-    let concrete_frame_addr = miframe.concrete_frame_addr;
-    let entry_py_pc = EntryPyPc::Py(miframe.orgpc as u32);
-    // SAFETY: both pointers were initialized at MIFrame
-    // construction time and outlive this call (TraceCtx and
-    // PyreSym are pinned by the surrounding tracing session).
-    // `&mut sym` is held only for the post-walk
-    // `last_exc_box`/`class_of_last_exc_is_const` writeback below;
-    // during the walk itself we re-borrow only specific sym fields
-    // (`registers_*`) into the `WalkContext` slices — the walker no
-    // longer takes a fresh `&mut PyreSym` inside any helper, so there
-    // is no Stacked-Borrows aliasing between WalkContext's slice
-    // borrows and a parallel sym reborrow (parity #2 — the prior
-    // `miframe_sym: Option<*mut PyreSym>` field has been removed).
-    let trace_ctx = unsafe { &mut *ctx_ptr };
-    let sym = unsafe { &mut *sym_ptr };
-
-    // Phase 7: this IS the full-body walk over the outer `sym.jitcode`,
-    // so guard snapshots can resolve a per-guard resume coordinate from
-    // `op_pc`.  Set `fbw_mode.snapshot_sym` for the walk's lifetime;
-    // `walker_capture_snapshot_for_last_guard` and
-    // `fbw_foriter_body_from_op_pc` read it.  This is the PRODUCTION
-    // default tracer: `trace.rs` enters `full_body_walk_trace` whenever
-    // `PYRE_FULL_BODY_WALK` is not explicitly `0` (the env gate defaults ON),
-    // so `fbw_mode.snapshot_sym` is non-null on every default-JIT
-    // run.  `PYRE_FULL_BODY_WALK=0` is the only opt-out
-    // (the transitional trait leg), which leaves the pointer null.
-    // Recover the portal EC red off `sym.frame` before the first opcode is
-    // dispatched (thus before any guard is recorded), caching it into
-    // `sym.execution_context`.  A bridge-from-guard sym whose ec color collides
-    // with a real frame slot is left `OpRef::NONE` by `setup_bridge_sym`, which
-    // defers recovery to `ensure_execution_context`.  The walker's
-    // snapshot-capture path runs `collect_outer_active_boxes` AFTER the guard,
-    // so recovering there would record the getfield after the guard that
-    // references it (use-before-def).  Seed here — the trait's pre-guard
-    // cache-once analog — so every guard snapshot reads a real EC OpRef.
-    seed_execution_context_for_walk(sym, trace_ctx);
-    seed_standing_exception_for_walk(sym, trace_ctx);
-
-    // RPython parity: `metainterp.last_exc_value` (pyjitpl.py:1695)
-    // is the standing exception OpRef. Walker's `WalkContext::last_exc_value`
-    // mirrors this as `Option<OpRef>` — `None` means "no active
-    // exception", matching RPython's `assert self.metainterp.last_exc_value`
-    // (pyjitpl.py:1702).
-    let initial_last_exc_value = if sym.last_exc_box.is_none() {
-        None
-    } else {
-        Some(sym.last_exc_box)
-    };
-
-    // PyPy `pyjitpl.py:171-176 MIFrame.__init__` analog: allocate
-    // fresh per-bank register vectors sized to `top_num_regs_* +
-    // top_constants_*.len()`.  This replaces the prior TODO
-    // that reused `sym.registers_r` (a Python locals/stack mirror,
-    // whose `[0]` slot is Python local 0) as the MIFrame register
-    // file.  The codewriter-compiled arm jitcode emits getfield
-    // chains rooted at `R[0] = handler = MIFrame self ptr`; the
-    // `argboxes_r` parameter supplies that handler ptr below via the
-    // `setup_call` analog.
-    let total_r = top_num_regs_r + top_constants_r.len();
-    let total_i = top_num_regs_i + top_constants_i.len();
-    let total_f = top_num_regs_f + top_constants_f.len();
-    let mut top_regs_r = vec![OpRef::NONE; total_r];
-    let mut top_regs_i = vec![OpRef::NONE; total_i];
-    let mut top_regs_f = vec![OpRef::NONE; total_f];
-    let mut top_concrete_r = vec![ConcreteValue::Null; total_r];
-    let mut top_concrete_i = vec![ConcreteValue::Null; total_i];
-
-    // PyPy `pyjitpl.py:98-119 copy_constants` analog: seed each
-    // constant into the upper slot range `[num_regs_*, total_*)`.
-    // `box_value` resolves these via `TraceCtx::constants` so
-    // downstream getfield chains see the constant's `Value::*`.
-    for (i, &v) in top_constants_i.iter().enumerate() {
-        top_regs_i[top_num_regs_i + i] = trace_ctx.const_int(v);
-        top_concrete_i[top_num_regs_i + i] = ConcreteValue::Int(v);
-    }
-    for (i, &v) in top_constants_r.iter().enumerate() {
-        top_regs_r[top_num_regs_r + i] = trace_ctx.const_ref(v);
-        if v != 0 {
-            top_concrete_r[top_num_regs_r + i] = ConcreteValue::Ref(v as pyre_object::PyObjectRef);
-        }
-    }
-    for (i, &v) in top_constants_f.iter().enumerate() {
-        top_regs_f[top_num_regs_f + i] = trace_ctx.const_float(v);
-    }
-
-    // PyPy `pyjitpl.py:188-200 setup_call(argboxes)` analog: write
-    // each argbox into the leading register slot.  The concrete
-    // shadow is derived from `box_value(box)` — for `ConstRef(ptr)`
-    // (the common case: argbox=miframe self ptr), this is
-    // `Some(Value::Ref(GcRef(ptr)))` resolved via the constant pool;
-    // for non-const argboxes it consults the `opref_concrete` stamp
-    // table.
-    //
-    // CodeRabbit Major (PR #89): reject oversized argbox lists up
-    // front instead of silently truncating with a per-loop `break`.
-    // The `_*_arity_mismatch` DispatchError shapes already exist for
-    // the inline-call paths (`InlineCall*ArityMismatch`); reuse them
-    // here so a caller/shape mismatch surfaces as a typed failure
-    // rather than a partially seeded frame.
-    if argboxes_r.len() > top_num_regs_r {
-        return Err(DispatchError::InlineCallArityMismatch {
-            pc: position,
-            provided: argboxes_r.len(),
-            callee_num_regs_r: top_num_regs_r,
-        });
-    }
-    if argboxes_i.len() > top_num_regs_i {
-        return Err(DispatchError::InlineCallIntArityMismatch {
-            pc: position,
-            provided: argboxes_i.len(),
-            callee_num_regs_i: top_num_regs_i,
-        });
-    }
-    if argboxes_f.len() > top_num_regs_f {
-        return Err(DispatchError::InlineCallFloatArityMismatch {
-            pc: position,
-            provided: argboxes_f.len(),
-            callee_num_regs_f: top_num_regs_f,
-        });
-    }
-    for (i, &box_ref) in argboxes_r.iter().enumerate() {
-        top_regs_r[i] = box_ref;
-        if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = trace_ctx.box_value(box_ref) {
-            top_concrete_r[i] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-        }
-    }
-    for (i, &box_ref) in argboxes_i.iter().enumerate() {
-        top_regs_i[i] = box_ref;
-        if let Some(majit_ir::Value::Int(v)) = trace_ctx.box_value(box_ref) {
-            top_concrete_i[i] = ConcreteValue::Int(v);
-        }
-    }
-    for (i, &box_ref) in argboxes_f.iter().enumerate() {
-        top_regs_f[i] = box_ref;
-    }
-
-    // Seed last_exc_value_concrete from
-    // sym.last_exc_value (the live PyObjectRef written by trait-side
-    // `seed_raised_exception` at `trace_opcode.rs:6646`).  Null when
-    // no active exception, matching `initial_last_exc_value == None`.
-    let initial_last_exc_value_concrete = if sym.last_exc_value.is_null() {
-        ConcreteValue::Null
-    } else {
-        ConcreteValue::Ref(sym.last_exc_value)
-    };
-
-    let result = {
-        let mut wc = WalkContext {
-            callee_shadow: None,
-            inline_callee_consts: None,
-            fbw_mode: FbwWalkMode {
-                snapshot_sym: sym_ptr,
-                current_exception_seed: (trace_ctx.is_bridge_trace && !sym.last_exc_box.is_none())
-                    .then_some(sym.last_exc_box),
-                current_exception_seed_concrete: if trace_ctx.is_bridge_trace {
-                    sym.last_exc_value
-                } else {
-                    pyre_object::PY_NULL
-                },
-                class_of_last_exc_is_const: sym.class_of_last_exc_is_const,
-                ..Default::default()
-            },
-            session,
-            registers_r: &mut top_regs_r,
-            registers_i: &mut top_regs_i,
-            registers_f: &mut top_regs_f,
-            concrete_registers_r: &mut top_concrete_r,
-            concrete_registers_i: &mut top_concrete_i,
-            descr_refs,
-            raw_descrs,
-            is_authoritative_executor,
-            // `dispatch_via_miframe` is the full-body walk entry
-            // (production tracer, diagnostic probe — the non-production
-            // roots are excluded from concrete execution by
-            // `is_authoritative_executor: false` instead).
-            is_full_body_walk: true,
-            trace_ctx,
-            done_with_this_frame_descr_ref,
-            done_with_this_frame_descr_int,
-            done_with_this_frame_descr_float,
-            done_with_this_frame_descr_void,
-            exit_frame_with_exception_descr_ref,
-            is_top_level,
-            sub_jitcode_lookup,
-            last_exc_value: initial_last_exc_value,
-            last_exc_value_concrete: initial_last_exc_value_concrete,
-            entry_py_pc,
-            outer_resume_marker_jit_pc: None,
-            outer_jitcode_index: 0,
-            outer_active_boxes: Vec::new(),
-            // This entry (test/fixture) hard-codes
-            // `outer_jitcode_index = 0` and an empty `outer_active_boxes`
-            // rather than seeding them from `sym.jitcode` /
-            // `collect_outer_active_boxes` like the retired per-opcode
-            // arm entry did.  A specialized
-            // `generated_store_subscr_value` guard captured via
-            // `walker_capture_snapshot_for_last_guard` would attach
-            // resume data pointing at the wrong frame, so keep
-            // STORE_SUBSCR specialization off on this entry.
-            store_subscr_fn_addr: None,
-            pending_guard_snapshot_error: None,
-            vstack_boxes: Vec::new(),
-            vstack_depth: 0,
-            vstack_cur_pypc: 0,
-            vstack_valid: false,
-            vstack_last_ref: OpRef::NONE,
-            vstack_reorder_ceiling: u32::MAX,
-            live_before_jit_pc: usize::MAX,
-            live_after_jit_pc: usize::MAX,
-        };
-        // #73: seed the walk-level operand-stack box mirror
-        // at entry.  The mirror is only enabled when the outer sym owns the
-        // virtualizable shadow (the production full-body loop trace) — the
-        // synthetic/test entries leave it disabled (`vstack_valid = false`).
-        // Seed at the FIRST-walked jitcode pc (`position`), not `entry_py_pc`,
-        // so the first `step_vstack_mirror` is a no-op (no spurious
-        // entry-boundary reconcile of the not-yet-executed first opcode).
-        // Pure side-data: the snapshot read stays LEGACY until a later slice
-        // makes it authoritative.
-        seed_vstack_mirror(&mut wc, sym, position);
-        let outcome = walk(jitcode_code, position, &mut wc);
-        // Read final last_exc_value before wc drops so the borrow
-        // checker can release sym for the writeback below.
-        let final_last_exc = wc.last_exc_value;
-        let final_class_of_last_exc_is_const = wc.fbw_mode.class_of_last_exc_is_const;
-        drop(wc);
-        // Full `sym.last_exc_*` state writeback parity.
-        //
-        // RPython `pyjitpl.py:1694-1696 opimpl_raise` sets THREE pieces
-        // of metainterp state when a raise fires:
-        //   self.metainterp.class_of_last_exc_is_const = True
-        //   self.metainterp.last_exc_value = exc_value_box.getref(rclass.OBJECTPTR)
-        //   self.metainterp.last_exc_box = exc_value_box
-        //
-        // Of these, the walker can produce:
-        //   - `last_exc_box`: the symbolic OpRef. Mirrored from
-        //     `wc.last_exc_value` (RPython's metainterp.last_exc_value
-        //     and last_exc_box are different fields — concrete pointer
-        //     vs Box — but the walker tracks only the symbolic one,
-        //     which lines up with `sym.last_exc_box`).
-        //   - `class_of_last_exc_is_const`: true after a raise/r or a
-        //     SubRaise routed into a catch handler. RPython sets this
-        //     in `opimpl_raise` (line 1694) AND `execute_ll_raised`
-        //     (pyjitpl.py:2752 with `constant=...` parameter — set
-        //     after GUARD_CLASS / GUARD_EXCEPTION). Walker's raise/r
-        //     arm always sets `wc.last_exc_value = Some(exc)` so
-        //     mirroring `Some` → const=true is RPython-orthodox.
-        //
-        // The walker CANNOT produce:
-        //   - `sym.last_exc_value` (concrete `PyObjectRef`): RPython
-        //     `exc_value_box.getref(rclass.OBJECTPTR)` reads the
-        //     concrete pointer at trace-recording time. The symbolic
-        //     walker has only OpRefs — concrete writeback is the
-        //     production tracer's responsibility (the trait-driven
-        //     `MIFrame::execute_opcode_step` path). This is a known
-        //     TODO (the walker is symbolic-only,
-        //     concrete state is fed by another path).
-        if let Some(exc) = final_last_exc {
-            sym.last_exc_box = exc;
-            sym.class_of_last_exc_is_const = final_class_of_last_exc_is_const;
-        }
-        outcome
-    };
-    result
-}
-
-/// Build the paused root portal frame for a multi-frame bridge-carrier
-/// sub-walk (#215 item 2 / P2 drain).  The root resumes at `root_pc` once the
-/// reconstructed deepest callee returns; the callee's in-callee guards must
-/// snapshot this frame on the walk framestack so a guard-failure resume
-/// rebuilds both Python frames.  Mirror of [`compute_inline_caller_frame`], but
-/// the root register banks come straight from the bridge-seeded `root_sym`
-/// rather than a live caller [`WalkContext`] (the root walk has not started —
-/// this resumes mid-flight).
-fn compute_bridge_root_parent_frame(
-    root_sym: &crate::state::PyreSym,
-    trace_ctx: &mut TraceCtx,
-    root_pc: usize,
-) -> Option<InlineParentFrame> {
-    if root_sym.jitcode.is_null() {
-        return None;
-    }
-    let jitcode_index = unsafe { (*root_sym.jitcode).index as u32 };
-    // `root_pc` (`resume_data.frames[0].pc`) is already the post-call resume
-    // point — the slot the inner frame's result lands in — so its Python
-    // coordinate is a direct backtranslation (no `semantic_fallthrough_pc`).
-    // This local remains needed by the outer-active-box collector; the paused
-    // parent frame itself carries `ParentResumeCoord::Backxlat(root_pc)`.
-    let root_py_pc = crate::state::backxlat_py_pc(jitcode_index as i32, root_pc as i32) as u32;
-    // Null the not-yet-produced call-result slot before collecting the active
-    // boxes (the reconstructed callee supplies it on `SubReturn`), mirroring
-    // `compute_inline_caller_frame`.  Operate on a clone so `root_sym` stays a
-    // shared borrow.
-    //
-    // `collect_outer_active_boxes` reads the Ref bank by abstract register
-    // color (`_get_list_of_active_boxes`, pyjitpl.py:216-233), so it needs the
-    // color-indexed `f.registers_r` (`consume_boxes`, resume.py:1055), NOT the
-    // slot-indexed semantic mirror `setup_bridge_sym` left in
-    // `sym.registers_r`.  The mirror leaves an operand live across the resumed
-    // call (e.g. `t1` in `return fib(n-1)+fib(n-2)`) at `OpRef::NONE` under its
-    // color, which resolves to a NULL const and aborts the second residual
-    // call.  Prefer the persisted color decode; fall back to `registers_r` for
-    // non-bridge callers (`bridge_registers_r == None`).
-    let mut regs_r = root_sym
-        .bridge_registers_r
-        .clone()
-        .unwrap_or_else(|| root_sym.registers_r.clone());
-    if let Some(result_color) = unsafe { &(*root_sym.jitcode).payload }
-        .result_color_trivia_for_jitcode_pc(root_pc)
-        .map(|c| c as usize)
-        .filter(|&c| c != u16::MAX as usize)
-    {
-        if result_color < regs_r.len() {
-            regs_r[result_color] = trace_ctx.const_ref(pyre_object::PY_NULL as i64);
-        }
-    }
-    let root_word = ((root_pc as i32) != majit_ir::resumedata::NO_JITCODE_PC
-        && (root_pc as i32) >= 0)
-        .then_some(root_pc);
-    let root_liveness_word = match root_word {
-        Some(w) => w as i32,
-        None => majit_ir::resumedata::NO_JITCODE_PC,
-    };
-    let boxes = collect_outer_active_boxes(
-        root_sym,
-        trace_ctx,
-        &root_sym.registers_i,
-        &regs_r,
-        &root_sym.registers_f,
-        jitcode_index,
-        root_py_pc,
-        None,
-        // Key the query off the same carried root-frame word the snapshot and
-        // decode side read from `frames[0].jitcode_pc`, so both resolve the
-        // identical liveness window.
-        root_liveness_word,
-        root_liveness_word,
-        OuterActiveBoxesEntryTwin::Trivia,
-        "bridge_root_parent",
-        None,
-        &[],
-    );
-    Some(InlineParentFrame {
-        jitcode_index,
-        call_jitcode_pc: None,
-        call_stack_overrides: Vec::new(),
-        resume_coord: ParentResumeCoord::Backxlat(root_pc),
-        // Parent-frame words are never branch-tagged; negative tags belong to
-        // a branch guard's own top-frame word.
-        resume_marker_jit_pc: root_word,
-        boxes,
-    })
-}
-
-/// Issue #215 item 2 (P2 drain, increment 2b): drive the reconstructed deepest
-/// callee frame of a multi-frame bridge as an INLINE SUB-WALK
-/// (`is_top_level = false`) rooted on the caller-visible portal `root_sym`.
-///
-/// The callee resumes at `entry` (its translated recipe Python pc) with
-/// its registers seeded by `argboxes_r` (portal reds + in-flight operand-stack
-/// temps from `setup_reconstructed_callee_frame`) and its locals carried in the
-/// already-emitted frame vable.  Because the walk is a sub-walk, the callee's
-/// `ref_return` surfaces `SubReturn { result }` (`pyjitpl.py:1688 finishframe`)
-/// instead of the top-level `Finish` that pyre's own-portal model rejects with
-/// `NonStandardVableFinishPortalUnsupported` — the original #215 item-2 wall.
-///
-/// The root portal is installed as `fbw_mode.snapshot_sym` and pushed onto the
-/// walk framestack for the sub-walk's lifetime, so an in-callee guard
-/// snapshots both the callee frame and the paused root
-/// (`walker_capture_multi_frame_inline_snapshot`).
-///
-/// Increment 2b-i (diagnostic): returns the sub-walk outcome; the caller logs it
-/// and aborts (trace discarded).  Threading `SubReturn` into the root operand
-/// stack + the root top-level walk forward to a terminator is increment 2b-ii.
-/// `None` signals a setup failure (terminal descrs unwired / no root frame).
-fn call_dst_reg_for_residual_return(code: &[u8], entry: usize) -> Option<usize> {
-    for op in crate::jitcode_runtime::decoded_ops(code) {
-        if op.next_pc == entry {
-            return (op.opname.starts_with("residual_call") && op.argcodes.ends_with(">r"))
-                .then(|| code.get(entry - 1).map(|&b| b as usize))
-                .flatten();
-        }
-    }
-    None
-}
-
-fn recipe_parent_frame_from_recipe(
-    ctx: &mut TraceCtx,
-    recipe: &majit_metainterp::ReconstructRecipe,
-    root_ec: *const pyre_interpreter::PyExecutionContext,
-) -> Option<InlineParentFrame> {
-    let pjc = crate::state::pyjitcode_for_jitcode_index(recipe.jitcode_index)?;
-    if !pjc.is_populated() || pjc.code_ptr.is_null() {
-        return None;
-    }
-    let entry =
-        if crate::state::frame_pc_is_resolved_offset_at(recipe.jitcode_index, recipe.jitcode_pc) {
-            recipe.jitcode_pc as usize
-        } else {
-            return None;
-        };
-    let call_jit_pc = crate::jitcode_runtime::decoded_ops(pjc.jitcode.code.as_slice())
-        .find(|op| op.next_pc == entry && op.opname.starts_with("residual_call"))
-        .map(|op| op.pc);
-    let resume_marker_jit_pc =
-        call_jit_pc.and_then(|pc| pjc.after_residual_marker_for_jitcode_pc(pc));
-
-    // Reconstruct this paused parent frame's vable + ec (the same
-    // `emit_new_pyframe_inline_with_params` the deepest-callee setup uses) so
-    // the paused-frame snapshot resolves the portal reds
-    // [frame, ec] (`interp_jit.py:67`) to real boxes rather than reading the
-    // slot-indexed `registers_r` at the portal-red color positions.  Only
-    // `pending.sym.frame` / `pending.sym.execution_context` are consumed here;
-    // the `argboxes_r` register seeding is for the forward drive, not the
-    // snapshot.
-    let (pending, _argboxes_r) =
-        crate::state::setup_reconstructed_callee_frame(ctx, recipe, root_ec, Vec::new())?;
-    let frame_box = pending.sym.frame;
-    let ec_box = pending.sym.execution_context;
-
-    let (frame_reg, ec_reg) = crate::state::portal_red_regs_at(recipe.jitcode_index);
-    let (frame_reg, ec_reg) = (u32::from(frame_reg), u32::from(ec_reg));
-    let sentinel = u32::from(u16::MAX);
-    let result_color = pjc
-        .result_color_trivia_for_jitcode_pc(recipe.jitcode_pc as usize)
-        .map(|c| c as usize)
-        .filter(|&c| c != u16::MAX as usize);
-
-    let banks = crate::state::frame_liveness_reg_indices_by_bank_from_pc(
-        recipe.jitcode_index,
-        recipe.jitcode_pc,
-    );
-    let stack_only = recipe.valuestackdepth.saturating_sub(recipe.nlocals);
-    let maps = crate::state::bridge_semantic_maps_from_pc(recipe.jitcode_index, recipe.jitcode_pc);
-    let null_ref = ctx.const_ref(pyre_object::PY_NULL as i64);
-    let mut boxes = Vec::with_capacity(banks.total_len());
-    for &color in &banks.int {
-        boxes.push(
-            recipe
-                .registers_i
-                .get(color as usize)
-                .copied()
-                .unwrap_or(OpRef::NONE),
-        );
-    }
-    // Ref bank, in liveness-color order — mirror the trait encoder
-    // `get_list_of_active_boxes` (trace_opcode.rs:1902-1957) box-for-box:
-    //   * the not-yet-produced call-result color is NULL-seeded (`in_a_call`);
-    //   * a force-alived portal-red SCRATCH color (no live semantic slot at this
-    //     pc) routes to the reconstructed frame's `frame`/`ec` box, NOT the
-    //     slot-indexed register file;
-    //   * every other live color reads its semantic `locals_cells_stack_w` slot
-    //     from the slot-indexed `registers_r` (the reconstruct decode).
-    for &color in &banks.ref_ {
-        let c = color as usize;
-        if result_color == Some(c) {
-            boxes.push(null_ref);
-            continue;
-        }
-        let semantic_idx = crate::state::semantic_ref_slot_for_reg_color(
-            recipe.nlocals,
-            stack_only,
-            &maps.pcdep_entries,
-            c,
-        );
-        let is_portal_red_scratch = semantic_idx.is_none()
-            && ((color == frame_reg && frame_reg != sentinel)
-                || (color == ec_reg && ec_reg != sentinel));
-        if is_portal_red_scratch {
-            boxes.push(if color == frame_reg {
-                frame_box
-            } else {
-                ec_box
-            });
-            continue;
-        }
-        let slot = semantic_idx.or_else(|| (c < recipe.valuestackdepth).then_some(c))?;
-        boxes.push(recipe.registers_r.get(slot).copied().unwrap_or(OpRef::NONE));
-    }
-    for &color in &banks.float {
-        boxes.push(
-            recipe
-                .registers_f
-                .get(color as usize)
-                .copied()
-                .unwrap_or(OpRef::NONE),
-        );
-    }
-    if boxes.iter().any(|b| b.is_none()) {
-        return None;
-    }
-
-    Some(InlineParentFrame {
-        jitcode_index: recipe.jitcode_index as u32,
-        call_jitcode_pc: call_jit_pc,
-        call_stack_overrides: Vec::new(),
-        // The recipe's resolved word was `backxlat_py_pc(jitcode_index,
-        // jitcode_pc)` by construction, exactly the bridge-root flavor.
-        resume_coord: ParentResumeCoord::Backxlat(recipe.jitcode_pc as usize),
-        resume_marker_jit_pc,
-        boxes,
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn drive_bridge_frame_subwalk(
-    ctx: &mut TraceCtx,
-    session: &std::cell::RefCell<WalkSession>,
-    root_sym: &crate::state::PyreSym,
-    root_pc: usize,
-    callee_pjc: &std::sync::Arc<crate::PyJitCode>,
-    callee_code_key: usize,
-    callee_w_globals: usize,
-    entry: usize,
-    argboxes_r: &[OpRef],
-    local_concretes: &[majit_ir::Value],
-    child_result: Option<OpRef>,
-    paused_parent_recipes: &[majit_metainterp::ReconstructRecipe],
-) -> Option<Result<(DispatchOutcome, usize), DispatchError>> {
-    use majit_metainterp::jitcode::RuntimeBhDescr;
-
-    // Terminal descrs off MetaInterpStaticData (mirror `dispatch_perfn_frame`).
-    let (done_void, done_int, done_ref, done_float, exit_exc_ref) = {
-        let sd = ctx.metainterp_sd();
-        match (
-            sd.done_with_this_frame_descr_void.clone(),
-            sd.done_with_this_frame_descr_int.clone(),
-            sd.done_with_this_frame_descr_ref.clone(),
-            sd.done_with_this_frame_descr_float.clone(),
-            sd.exit_frame_with_exception_descr_ref.clone(),
-        ) {
-            (Some(v), Some(i), Some(r), Some(f), Some(e)) => (v, i, r, f, e),
-            _ => return None,
-        }
-    };
-
-    // Per-fn descr pool + sub-jitcode lookup off the callee body's own runtime
-    // pool (mirror `dispatch_perfn_frame`).  `callee_pjc` is an `Arc` that
-    // outlives the walk, so extend the descr-slice borrow to `'static` for the
-    // `'static`-bodied `SubJitCodeBody` lookup.
-    let perfn_descrs: &'static [RuntimeBhDescr] =
-        unsafe { &*(callee_pjc.jitcode.exec.descrs.as_slice() as *const [RuntimeBhDescr]) };
-    let perfn_descr_refs: Vec<majit_ir::DescrRef> = perfn_descrs
-        .iter()
-        .enumerate()
-        .map(|(i, d)| match d {
-            RuntimeBhDescr::Descr(bh) => crate::descr::make_descr_from_bh(bh),
-            RuntimeBhDescr::JitCode(_)
-            | RuntimeBhDescr::Call(_)
-            | RuntimeBhDescr::AssemblerToken(_) => crate::descr::make_jitcode_descr(i),
-        })
-        .collect();
-    let sub_jitcode_lookup = |idx: usize| -> Option<SubJitCodeBody> {
-        perfn_descrs
-            .get(idx)
-            .and_then(|d| d.as_jitcode())
-            .map(|jc| SubJitCodeBody {
-                code: jc.code.as_slice(),
-                num_regs_r: jc.num_regs_r() as usize,
-                num_regs_i: jc.num_regs_i() as usize,
-                num_regs_f: jc.num_regs_f() as usize,
-                constants_i: jc.constants_i.as_slice(),
-                constants_r: jc.constants_r.as_slice(),
-                constants_f: jc.constants_f.as_slice(),
-            })
-    };
-
-    // Allocate the callee register banks sized to `num_regs_* + constants_*`,
-    // seed the constant pool into the upper slots and `argboxes_r` into the
-    // leading slots (mirror `dispatch_via_miframe`).
-    let jc = &callee_pjc.jitcode;
-    let num_regs_r = jc.num_regs_r() as usize;
-    let num_regs_i = jc.num_regs_i() as usize;
-    let num_regs_f = jc.num_regs_f() as usize;
-    let total_r = num_regs_r + jc.constants_r.len();
-    let total_i = num_regs_i + jc.constants_i.len();
-    let total_f = num_regs_f + jc.constants_f.len();
-    let mut regs_r = vec![OpRef::NONE; total_r];
-    let mut regs_i = vec![OpRef::NONE; total_i];
-    let mut regs_f = vec![OpRef::NONE; total_f];
-    let mut concrete_r = vec![ConcreteValue::Null; total_r];
-    let mut concrete_i = vec![ConcreteValue::Null; total_i];
-    for (i, &v) in jc.constants_i.iter().enumerate() {
-        regs_i[num_regs_i + i] = ctx.const_int(v);
-        concrete_i[num_regs_i + i] = ConcreteValue::Int(v);
-    }
-    for (i, &v) in jc.constants_r.iter().enumerate() {
-        regs_r[num_regs_r + i] = ctx.const_ref(v);
-        if v != 0 {
-            concrete_r[num_regs_r + i] = ConcreteValue::Ref(v as pyre_object::PyObjectRef);
-        }
-    }
-    for (i, &v) in jc.constants_f.iter().enumerate() {
-        regs_f[num_regs_f + i] = ctx.const_float(v);
-    }
-    if argboxes_r.len() > num_regs_r {
-        return Some(Err(DispatchError::InlineCallArityMismatch {
-            pc: entry,
-            provided: argboxes_r.len(),
-            callee_num_regs_r: num_regs_r,
-        }));
-    }
-    for (i, &box_ref) in argboxes_r.iter().enumerate() {
-        regs_r[i] = box_ref;
-        if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = ctx.box_value(box_ref) {
-            concrete_r[i] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-        }
-    }
-    if let Some(result) = child_result {
-        let call_dst_reg = call_dst_reg_for_residual_return(jc.code.as_slice(), entry)?;
-        if call_dst_reg >= regs_r.len() {
-            return Some(Err(DispatchError::InlineCallArityMismatch {
-                pc: entry,
-                provided: call_dst_reg + 1,
-                callee_num_regs_r: regs_r.len(),
-            }));
-        }
-        regs_r[call_dst_reg] = result;
-        if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = ctx.box_value(result) {
-            concrete_r[call_dst_reg] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-        }
-    }
-
-    // Paused root portal frame for the multi-frame guard snapshot.
-    let root_frame = compute_bridge_root_parent_frame(root_sym, ctx, root_pc)?;
-    let outer_jitcode_index = root_frame.jitcode_index;
-    let outer_active_boxes = root_frame.boxes.clone();
-
-    let callee_code = jc.code.as_slice();
-    let lookup_ref: &SubJitCodeLookup = &sub_jitcode_lookup;
-    let consts = InlineCalleeConsts {
-        w_globals: callee_w_globals,
-        w_code: callee_code_key,
-    };
-
-    // Install the ROOT sym as the snapshot sym (NOT the callee's) so in-callee
-    // guards snapshot the paused root.
-    let root_sym_ptr = root_sym as *const crate::state::PyreSym;
-
-    let mut parent_guards = Vec::new();
-    let mut parent_for_current = root_frame.clone();
-    for parent_recipe in paused_parent_recipes {
-        let guard_parent = parent_for_current.clone();
-        parent_guards.push(InlineFrameGuard::enter(
-            session,
-            parent_recipe.code_ptr as usize,
-            Some(guard_parent),
-        ));
-        parent_for_current = recipe_parent_frame_from_recipe(
-            ctx,
-            parent_recipe,
-            root_sym.concrete_execution_context,
-        )?;
-    }
-
-    let outcome = {
-        let mut sub_wc = WalkContext {
-            callee_shadow: Some(Default::default()),
-            inline_callee_consts: Some(consts),
-            fbw_mode: FbwWalkMode {
-                snapshot_sym: root_sym_ptr,
-                inline_subwalk: true,
-                carrier_resume: true,
-                current_exception_seed: (!root_sym.last_exc_box.is_none())
-                    .then_some(root_sym.last_exc_box),
-                current_exception_seed_concrete: root_sym.last_exc_value,
-                class_of_last_exc_is_const: root_sym.class_of_last_exc_is_const,
-            },
-            session,
-            registers_r: &mut regs_r,
-            registers_i: &mut regs_i,
-            registers_f: &mut regs_f,
-            concrete_registers_r: &mut concrete_r,
-            concrete_registers_i: &mut concrete_i,
-            descr_refs: &perfn_descr_refs,
-            raw_descrs: RawDescrPool::PerFn(perfn_descrs),
-            // The carrier sub-walk IS the bridge-resume metainterp: after a
-            // guard failure `handle_guard_failure` rebuilds the frame state and
-            // drives it forward through the SAME `self.interpret()` the initial
-            // trace uses (`pyjitpl.py:2937 _handle_guard_failure` →
-            // `prepare_resume_from_failure` → `interpret`, cf.
-            // `_compile_and_run_once:2899`).  There is no second-class executor
-            // mode: the resume walk concrete-executes every residual call
-            // (`do_residual_call` → `execute_varargs`, `pyjitpl.py:1995`) exactly
-            // like the initial trace, which is what lets a nested self-recursive
-            // call fold to a live `CALL_ASSEMBLER`.  The residual it reaches was
-            // never run pre-deopt (the deopt cut the trace there), so this is its
-            // first and only concrete execution.
-            is_authoritative_executor: true,
-            is_full_body_walk: true,
-            store_subscr_fn_addr: None,
-            pending_guard_snapshot_error: None,
-            vstack_boxes: Vec::new(),
-            vstack_depth: 0,
-            vstack_cur_pypc: 0,
-            vstack_valid: false,
-            vstack_last_ref: OpRef::NONE,
-            vstack_reorder_ceiling: u32::MAX,
-            live_before_jit_pc: usize::MAX,
-            live_after_jit_pc: usize::MAX,
-            trace_ctx: ctx,
-            done_with_this_frame_descr_ref: done_ref,
-            done_with_this_frame_descr_int: done_int,
-            done_with_this_frame_descr_float: done_float,
-            done_with_this_frame_descr_void: done_void,
-            exit_frame_with_exception_descr_ref: exit_exc_ref,
-            is_top_level: false,
-            sub_jitcode_lookup: lookup_ref,
-            last_exc_value: None,
-            last_exc_value_concrete: ConcreteValue::Null,
-            // The outer Python frame is the root, paused at `root_pc`.
-            entry_py_pc: EntryPyPc::Jit(root_pc),
-            outer_resume_marker_jit_pc: root_frame.resume_marker_jit_pc,
-            outer_jitcode_index,
-            outer_active_boxes,
-        };
-        let _inline_frame =
-            InlineFrameGuard::enter(session, callee_code_key, Some(parent_for_current));
-        // Nested self-recursive calls inside the resumed callee fold straight to
-        // a recursive-portal CALL_ASSEMBLER (the bridge is the deopt
-        // continuation, not a fresh unroll).
-        // Seed the reconstructed callee's local slot concretes into its frame-
-        // owned shadow.  The resume is mid-body,
-        // so the locals were stored to the frame vable before the guard fired;
-        // the map is empty until seeded.  A concrete local lets a callee
-        // `getarrayitem_vable(frame, slot)` read fold to its value, so a nested
-        // self-recursive call's int arg is known (`arg_is_int`) and the call
-        // folds to `CALL_ASSEMBLER` instead of declining.
-        for (slot, &v) in local_concretes.iter().enumerate() {
-            sub_wc.callee_shadow.as_mut().unwrap().set_concrete(
-                callee_pjc.metadata.portal_frame_reg,
-                slot as i64,
-                v,
-            );
-        }
-        walk(callee_code, entry, &mut sub_wc)
-    };
-    drop(parent_guards);
-    Some(outcome)
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn drive_bridge_carrier_subwalk(
-    ctx: &mut TraceCtx,
-    session: &std::cell::RefCell<WalkSession>,
-    root_sym: &crate::state::PyreSym,
-    root_pc: usize,
-    callee_pjc: &std::sync::Arc<crate::PyJitCode>,
-    callee_code_key: usize,
-    callee_w_globals: usize,
-    entry: usize,
-    argboxes_r: &[OpRef],
-    local_concretes: &[majit_ir::Value],
-    paused_parent_recipes: &[majit_metainterp::ReconstructRecipe],
-) -> Option<Result<(DispatchOutcome, usize), DispatchError>> {
-    drive_bridge_frame_subwalk(
-        ctx,
-        session,
-        root_sym,
-        root_pc,
-        callee_pjc,
-        callee_code_key,
-        callee_w_globals,
-        entry,
-        argboxes_r,
-        local_concretes,
-        None,
-        paused_parent_recipes,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn drive_bridge_middle_frame(
-    ctx: &mut TraceCtx,
-    session: &std::cell::RefCell<WalkSession>,
-    root_sym: &crate::state::PyreSym,
-    root_pc: usize,
-    middle_pjc: &std::sync::Arc<crate::PyJitCode>,
-    middle_code_key: usize,
-    middle_w_globals: usize,
-    entry: usize,
-    argboxes_r: &[OpRef],
-    local_concretes: &[majit_ir::Value],
-    child_result: OpRef,
-) -> Option<Result<(DispatchOutcome, usize), DispatchError>> {
-    drive_bridge_frame_subwalk(
-        ctx,
-        session,
-        root_sym,
-        root_pc,
-        middle_pjc,
-        middle_code_key,
-        middle_w_globals,
-        entry,
-        argboxes_r,
-        local_concretes,
-        Some(child_result),
-        &[],
-    )
-}
-
-/// #41 continuous cross-frame walk: after the deepest callee sub-walk
-/// ([`drive_bridge_carrier_subwalk`]) returns the callee result, continue the
-/// OUTER (root portal) frame forward from its resume pc, APPENDING to the same
-/// `ctx` (no cut, no fresh `run_perfn_walk` — that resets the trace and
-/// discards the sub-walk's live `CALL_ASSEMBLER` continuation).
-///
-/// The outer resumes AFTER its residual call (`entry` = the jitcode pc of the
-/// `live/` marker following the call), so the callee `result` is delivered into
-/// the outer's call-dst register (`call_dst_reg`) — the physical slot the call
-/// op wrote, i.e. `make_result_of_lastop` (`pyjitpl.py:258-275`), NOT a resume
-/// color. The frame vable identity is seeded at `frame_reg` (the standard
-/// virtualizable box) so the outer's local reads (`getarrayitem_vable`) resolve
-/// off the paused root frame. `is_top_level=true`: the outer IS the portal, so
-/// its `*_return` surfaces the portal `Terminate`/finish, not a `SubReturn`.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn drive_outer_frame_continuation(
-    ctx: &mut TraceCtx,
-    session: &std::cell::RefCell<WalkSession>,
-    root_sym: &crate::state::PyreSym,
-    root_pjc: &std::sync::Arc<crate::PyJitCode>,
-    root_code_key: usize,
-    root_w_globals: usize,
-    root_pc: usize,
-    entry: usize,
-    frame_box: OpRef,
-    frame_reg: usize,
-    result: OpRef,
-    call_dst_reg: usize,
-) -> Option<Result<(DispatchOutcome, usize), DispatchError>> {
-    use majit_metainterp::jitcode::RuntimeBhDescr;
-
-    let (done_void, done_int, done_ref, done_float, exit_exc_ref) = {
-        let sd = ctx.metainterp_sd();
-        match (
-            sd.done_with_this_frame_descr_void.clone(),
-            sd.done_with_this_frame_descr_int.clone(),
-            sd.done_with_this_frame_descr_ref.clone(),
-            sd.done_with_this_frame_descr_float.clone(),
-            sd.exit_frame_with_exception_descr_ref.clone(),
-        ) {
-            (Some(v), Some(i), Some(r), Some(f), Some(e)) => (v, i, r, f, e),
-            _ => return None,
-        }
-    };
-
-    let perfn_descrs: &'static [RuntimeBhDescr] =
-        unsafe { &*(root_pjc.jitcode.exec.descrs.as_slice() as *const [RuntimeBhDescr]) };
-    let perfn_descr_refs: Vec<majit_ir::DescrRef> = perfn_descrs
-        .iter()
-        .enumerate()
-        .map(|(i, d)| match d {
-            RuntimeBhDescr::Descr(bh) => crate::descr::make_descr_from_bh(bh),
-            RuntimeBhDescr::JitCode(_)
-            | RuntimeBhDescr::Call(_)
-            | RuntimeBhDescr::AssemblerToken(_) => crate::descr::make_jitcode_descr(i),
-        })
-        .collect();
-    let sub_jitcode_lookup = |idx: usize| -> Option<SubJitCodeBody> {
-        perfn_descrs
-            .get(idx)
-            .and_then(|d| d.as_jitcode())
-            .map(|jc| SubJitCodeBody {
-                code: jc.code.as_slice(),
-                num_regs_r: jc.num_regs_r() as usize,
-                num_regs_i: jc.num_regs_i() as usize,
-                num_regs_f: jc.num_regs_f() as usize,
-                constants_i: jc.constants_i.as_slice(),
-                constants_r: jc.constants_r.as_slice(),
-                constants_f: jc.constants_f.as_slice(),
-            })
-    };
-
-    let jc = &root_pjc.jitcode;
-    let num_regs_r = jc.num_regs_r() as usize;
-    let num_regs_i = jc.num_regs_i() as usize;
-    let num_regs_f = jc.num_regs_f() as usize;
-    let total_r = num_regs_r + jc.constants_r.len();
-    let total_i = num_regs_i + jc.constants_i.len();
-    let total_f = num_regs_f + jc.constants_f.len();
-    let mut regs_r = vec![OpRef::NONE; total_r];
-    let mut regs_i = vec![OpRef::NONE; total_i];
-    let mut regs_f = vec![OpRef::NONE; total_f];
-    let mut concrete_r = vec![ConcreteValue::Null; total_r];
-    let mut concrete_i = vec![ConcreteValue::Null; total_i];
-    for (i, &v) in jc.constants_i.iter().enumerate() {
-        regs_i[num_regs_i + i] = ctx.const_int(v);
-        concrete_i[num_regs_i + i] = ConcreteValue::Int(v);
-    }
-    for (i, &v) in jc.constants_r.iter().enumerate() {
-        regs_r[num_regs_r + i] = ctx.const_ref(v);
-        if v != 0 {
-            concrete_r[num_regs_r + i] = ConcreteValue::Ref(v as pyre_object::PyObjectRef);
-        }
-    }
-    for (i, &v) in jc.constants_f.iter().enumerate() {
-        regs_f[num_regs_f + i] = ctx.const_float(v);
-    }
-
-    let root_frame = compute_bridge_root_parent_frame(root_sym, ctx, root_pc)?;
-    let outer_jitcode_index = root_frame.jitcode_index;
-    let outer_active_boxes = root_frame.boxes.clone();
-
-    // Seed the full live outer-frame register file, matching `consume_boxes`
-    // (resume.py:1054-1055 + `_callback_i/_r/_f`) which fills every live
-    // register color of `framestack[-1]` from the resume numbering.  Without
-    // this only `frame_reg`/`call_dst_reg` are bound: an operand live across
-    // the resumed call (e.g. the first result of `return fib(n-1)+fib(n-2)`)
-    // stays `OpRef::NONE`, and the second residual call aborts with
-    // `ResidualCallArgUnbound`.
-    //
-    // `root_frame.boxes` already holds each live register's resolved box in
-    // `banks.int ++ banks.ref_ ++ banks.float` liveness order
-    // (`collect_outer_active_boxes`), applying the color->slot inversion the
-    // Ref bank needs.  Re-query the same (deterministic) liveness banks to
-    // recover each box's register color and scatter it into the color-indexed
-    // walker banks.  Only live colors are touched; dead slots stay
-    // `OpRef::NONE` so a later guard snapshot cannot capture a stale operand.
-    {
-        let root_word = ((root_pc as i32) != majit_ir::resumedata::NO_JITCODE_PC
-            && (root_pc as i32) >= 0)
-            .then_some(root_pc);
-        // Mirror `compute_bridge_root_parent_frame` so scatter reads the banks in collection order.
-        let banks = crate::state::frame_liveness_reg_indices_by_bank_at_with_jitcode_pc(
-            outer_jitcode_index as i32,
-            match root_word {
-                Some(w) => w as i32,
-                None => majit_ir::resumedata::NO_JITCODE_PC,
-            },
-        );
-        let mut cursor = 0usize;
-        for &color in &banks.int {
-            let v = outer_active_boxes
-                .get(cursor)
-                .copied()
-                .unwrap_or(OpRef::NONE);
-            cursor += 1;
-            let c = color as usize;
-            if c < regs_i.len() && v != OpRef::NONE {
-                regs_i[c] = v;
-                if let Some(majit_ir::Value::Int(n)) = ctx.box_value(v) {
-                    concrete_i[c] = ConcreteValue::Int(n);
-                }
-            }
-        }
-        for &color in &banks.ref_ {
-            let v = outer_active_boxes
-                .get(cursor)
-                .copied()
-                .unwrap_or(OpRef::NONE);
-            cursor += 1;
-            let c = color as usize;
-            if c < regs_r.len() && v != OpRef::NONE {
-                regs_r[c] = v;
-                if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = ctx.box_value(v) {
-                    if ptr != 0 && ptr != usize::MAX {
-                        concrete_r[c] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-                    }
-                }
-            }
-        }
-        for &color in &banks.float {
-            let v = outer_active_boxes
-                .get(cursor)
-                .copied()
-                .unwrap_or(OpRef::NONE);
-            cursor += 1;
-            let c = color as usize;
-            if c < regs_f.len() && v != OpRef::NONE {
-                regs_f[c] = v;
-            }
-        }
-    }
-
-    // Seed the standard virtualizable identity (frame) so the outer's vable
-    // reads hit the standard fast path, and the delivered callee result into the
-    // outer's call-dst register (`make_result_of_lastop`).  These overwrite the
-    // scattered snapshot values (the call-dst slot was nulled in
-    // `compute_bridge_root_parent_frame` for the not-yet-produced result).
-    if frame_reg < regs_r.len() {
-        regs_r[frame_reg] = frame_box;
-        if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = ctx.box_value(frame_box) {
-            concrete_r[frame_reg] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-        }
-    }
-    if call_dst_reg < regs_r.len() {
-        regs_r[call_dst_reg] = result;
-        if let Some(majit_ir::Value::Ref(majit_ir::GcRef(ptr))) = ctx.box_value(result) {
-            concrete_r[call_dst_reg] = ConcreteValue::Ref(ptr as pyre_object::PyObjectRef);
-        }
-    }
-
-    let root_code = jc.code.as_slice();
-    let lookup_ref: &SubJitCodeLookup = &sub_jitcode_lookup;
-    let consts = InlineCalleeConsts {
-        w_globals: root_w_globals,
-        w_code: root_code_key,
-    };
-
-    let root_sym_ptr = root_sym as *const crate::state::PyreSym;
-
-    let outcome = {
-        let mut outer_wc = WalkContext {
-            callee_shadow: None,
-            inline_callee_consts: Some(consts),
-            fbw_mode: FbwWalkMode {
-                snapshot_sym: root_sym_ptr,
-                inline_subwalk: true,
-                carrier_resume: true,
-                current_exception_seed: (!root_sym.last_exc_box.is_none())
-                    .then_some(root_sym.last_exc_box),
-                current_exception_seed_concrete: root_sym.last_exc_value,
-                class_of_last_exc_is_const: root_sym.class_of_last_exc_is_const,
-            },
-            session,
-            registers_r: &mut regs_r,
-            registers_i: &mut regs_i,
-            registers_f: &mut regs_f,
-            concrete_registers_r: &mut concrete_r,
-            concrete_registers_i: &mut concrete_i,
-            descr_refs: &perfn_descr_refs,
-            raw_descrs: RawDescrPool::PerFn(perfn_descrs),
-            is_authoritative_executor: true,
-            is_full_body_walk: true,
-            store_subscr_fn_addr: None,
-            pending_guard_snapshot_error: None,
-            vstack_boxes: Vec::new(),
-            vstack_depth: 0,
-            vstack_cur_pypc: 0,
-            vstack_valid: false,
-            vstack_last_ref: OpRef::NONE,
-            vstack_reorder_ceiling: u32::MAX,
-            live_before_jit_pc: usize::MAX,
-            live_after_jit_pc: usize::MAX,
-            trace_ctx: ctx,
-            done_with_this_frame_descr_ref: done_ref,
-            done_with_this_frame_descr_int: done_int,
-            done_with_this_frame_descr_float: done_float,
-            done_with_this_frame_descr_void: done_void,
-            exit_frame_with_exception_descr_ref: exit_exc_ref,
-            is_top_level: true,
-            sub_jitcode_lookup: lookup_ref,
-            last_exc_value: None,
-            last_exc_value_concrete: ConcreteValue::Null,
-            entry_py_pc: EntryPyPc::Jit(root_pc),
-            outer_resume_marker_jit_pc: root_frame.resume_marker_jit_pc,
-            outer_jitcode_index,
-            outer_active_boxes,
-        };
-        let _inline_frame = InlineFrameGuard::enter(session, root_code_key, Some(root_frame));
-        // The outer's own second self-recursive call folds to a live
-        // recursive-portal CALL_ASSEMBLER (same as the callee's), not a fresh
-        // unroll.
-        walk(root_code, entry, &mut outer_wc)
-    };
-    Some(outcome)
-}
-
-/// `getarrayitem_gc_<i|r|f>/rid>X` handler. Operand layout `rid>X`:
-/// 1B r-reg(array) + 1B i-reg(index) + 2B descr + 1B X-dst.
-///
-/// RPython parity: `pyjitpl.py:639-673 _do_getarrayitem_gc_any`:
-///
-///   tobox = heapcache.getarrayitem(arraybox, indexbox, arraydescr)
-///   if tobox: return tobox        # cache hit, no IR (recording-only)
-///   resop = self.execute_with_descr(op, arraydescr, arraybox, indexbox)
-///   heapcache.getarrayitem_now_known(arraybox, indexbox, resop, arraydescr)
-///   return resop
-///
-/// `opcode` is one of `GetarrayitemGc{I,R,F}`; `dst_bank` selects the
-/// result bank (`'i'`/`'r'`/`'f'`) the walker writes back into.  The
-/// index operand is always int-classified, so it is decoded from the
-/// `i` register bank.
-fn getarrayitem_gc_via_heapcache(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    opcode: OpCode,
-    dst_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    let array = read_ref_reg(code, op, 0, ctx)?;
-    let index = read_int_reg(code, op, 1, ctx)?;
-    let descr = read_descr(code, op, 2, ctx)?;
-    let descr_index = descr.index();
-
-    let result = if let Some(cached) =
-        ctx.trace_ctx
-            .heapcache_getarrayitem(array, index, descr_index)
-    {
-        // pyjitpl.py:639-673 `_do_getarrayitem_gc_any` cache hit:
-        //   tobox = heapcache.getarrayitem(...)
-        //   if tobox:
-        //       profiler.count_ops(rop.GETARRAYITEM_GC_I, HEAPCACHED_OPS)
-        //       return tobox
-        // RPython hardcodes `GETARRAYITEM_GC_I` regardless of the
-        // recorded `typ` ('i' / 'r' / 'f'); pyre matches the hardcode
-        // for profiling parity.
-        ctx.trace_ctx.profiler().count_ops(
-            OpCode::GetarrayitemGcI,
-            majit_metainterp::counters::HEAPCACHED_OPS,
-        );
-        cached
-    } else {
-        let resbox = ctx
-            .trace_ctx
-            .record_op_with_descr(opcode, &[array, index], descr.clone());
-        // Box.value parity: `box_value` exposes the resolution chain
-        // PyPy reads off `arraybox.getref_base()` / `indexbox.getint()`
-        // (`rpython/jit/metainterp/executor.py:117`).  Any operand
-        // whose Box.value is known unblocks `array_sanity_load`, not
-        // just Const-pool entries (`pyjitpl.py:648-666 resbox =
-        // execute_with_descr(...); getarrayitem_now_known(...)`
-        // parity).
-        let load_type = match opcode {
-            OpCode::GetarrayitemGcI | OpCode::GetarrayitemGcPureI => Some(majit_ir::Type::Int),
-            OpCode::GetarrayitemGcR | OpCode::GetarrayitemGcPureR => Some(majit_ir::Type::Ref),
-            OpCode::GetarrayitemGcF | OpCode::GetarrayitemGcPureF => Some(majit_ir::Type::Float),
-            _ => None,
-        };
-        let live_value = if let (
-            Some(ty),
-            Some(majit_ir::Value::Ref(array_ref)),
-            Some(majit_ir::Value::Int(index_value)),
-        ) = (
-            load_type,
-            ctx.trace_ctx.box_value(array),
-            ctx.trace_ctx.box_value(index),
-        ) {
-            let array_ptr = array_ref.0 as i64;
-            if array_ptr != usize::MAX as i64 && array_ptr != 0 {
-                ctx.trace_ctx
-                    .array_sanity_load(array_ptr, index_value, &descr, ty)
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-        // Stamp the loaded value as Box.value of the recorded result
-        // (RPython `Box(value)` constructor analog) so subsequent
-        // consumers see the runtime concrete instead of the
-        // GcRef(usize::MAX) sentinel.
-        if let Some(live_value) = live_value {
-            ctx.trace_ctx.set_opref_concrete(resbox, live_value);
-        }
-        ctx.trace_ctx
-            .heapcache_getarrayitem_now_known(array, index, descr_index, resbox);
-        resbox
-    };
-
-    let dst = code[op.pc + 5] as usize;
-    // concrete_of_opref derivation: derive shadow concrete from the recorded result's
-    // `concrete_of_opref` entry instead of inventing Null.  Constant
-    // arraybox + constant index hits land in `constants.get_value`;
-    // virtualizable hits surface via `standard_virtualizable_box`;
-    // `set_opref_concrete` stamps from upstream `binop_int_record`
-    // flow back here too.  Null fallback preserves the prior contract.
-    let concrete_for_shadow = concrete_from_recorded_opref(ctx, result);
-    match dst_bank {
-        'i' => {
-            write_int_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'r' => {
-            write_ref_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'f' => {
-            let len = ctx.registers_f.len();
-            let slot = ctx
-                .registers_f
-                .get_mut(dst)
-                .ok_or(DispatchError::RegisterOutOfRange {
-                    pc: op.pc,
-                    reg: dst,
-                    len,
-                    bank: "f",
-                })?;
-            *slot = result;
-        }
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `setarrayitem_gc_<i|r|f>/ri{i,r,f}d` handler. Operand layout per
-/// `bhimpl_setarrayitem_gc_{i,r,f}(cpu, array, index, newvalue,
-/// arraydescr)` (`blackhole.py:1351-1359`):
-/// 1B r-reg(array) + 1B i-reg(index) + 1B {i,r,f}-reg(newvalue) + 2B descr.
-///
-/// RPython parity: `pyjitpl.py:736-744 _opimpl_setarrayitem_gc_any`
-/// dispatches through `metainterp.execute_setarrayitem_gc(arraydescr,
-/// arraybox, indexbox, itembox)` — RPython's wrapper records
-/// `rop.SETARRAYITEM_GC` and updates the heapcache via
-/// `setarrayitem`.
-///
-/// No skip-on-redundant short-circuit (matches RPython —
-/// `_opimpl_setarrayitem_gc_any` has no `if cached == value: return`,
-/// because `heapcache.setarrayitem` already handles aliasing
-/// invalidation at the right granularity).
-///
-/// `value_bank` selects the newvalue register source: `'i'` /
-/// `'r'` / `'f'`.
-fn setarrayitem_gc_via_heapcache(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    value_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    let array = read_ref_reg(code, op, 0, ctx)?;
-    let index = read_int_reg(code, op, 1, ctx)?;
-    // Operand layout `<r><i><v>d`: r-reg(array) + i-reg(index) + v(value)
-    // + 2B descr-index.  For the `c`-coded short form
-    // (`setarrayitem_gc_i/ricd`) the value byte is an inline signed
-    // constant (`signedord`, `blackhole.py:123`) read as a `ConstInt`
-    // box instead of an `i`-register slot, mirroring `setfield_gc_i/rcd`.
-    let value = match value_bank {
-        'i' => read_int_reg(code, op, 2, ctx)?,
-        'r' => read_ref_reg(code, op, 2, ctx)?,
-        'f' => read_float_reg(code, op, 2, ctx)?,
-        'c' => OpRef::ConstInt(code[op.pc + 3] as i8 as i64),
-        _ => unreachable!("value_bank must be 'i', 'r', 'f' or 'c'"),
-    };
-    let descr = read_descr(code, op, 3, ctx)?;
-    let descr_index = descr.index();
-
-    ctx.trace_ctx
-        .record_op_with_descr(OpCode::SetarrayitemGc, &[array, index, value], descr);
-    // `upd.setarrayitem(valuebox)` (heapcache.py:142) parity — the
-    // cache stores the Box identity (`value` OpRef); cache-hit
-    // readers fetch the intrinsic value via `box_value(cached)` at
-    // hit time.
-    ctx.trace_ctx
-        .heapcache_setarrayitem(array, index, descr_index, value);
-    walker_fill_materialized_array(ctx, array, index, value);
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// Walker virtual-force fill — companion to the `NEW_ARRAY_CLEAR`
-/// materialization in the `new_array_clear` handler (module-global
-/// fresh-container off-by-one fix). When `array` is a still-unescaped block we
-/// materialized to a concrete GC `ItemsBlock` at NEW_ARRAY_CLEAR, write the
-/// concrete element into it so a later BUILD_LIST / BUILD_TUPLE residual reads
-/// a complete block during the walk. If the element value (not a ref) or the
-/// index has no known concrete, the block cannot be completed, so revert the
-/// array to the no-concrete sentinel (`Ref(usize::MAX)`): the residual then
-/// declines and the void store aborts, exactly as without materialization.
-///
-/// A real (already-escaped) array store — whose `array` operand is a
-/// `GetfieldGcR` load of a live container's items block, not a fresh
-/// allocation — is left untouched because `is_unescaped(array)` is false, so
-/// the runtime trace's own SETARRAYITEM is never duplicated eagerly here.
-fn walker_fill_materialized_array(
-    ctx: &mut WalkContext<'_, '_>,
-    array: OpRef,
-    index: OpRef,
-    value: OpRef,
-) {
-    // heapcache.py:493 is_unescaped — only a fresh, not-yet-escaped allocation
-    // is a materialization candidate; a loaded (escaped) array is a real store.
-    if !ctx.trace_ctx.heap_cache().is_unescaped(array) {
-        return;
-    }
-    let block = match ctx.trace_ctx.box_value(array) {
-        Some(majit_ir::Value::Ref(r)) if r != majit_ir::GcRef::NO_CONCRETE && r.as_usize() != 0 => {
-            r.as_usize() as *mut pyre_object::object_array::ItemsBlock
-        }
-        // No stamped concrete → not materialized (gate off / non-ref / non-const).
-        _ => return,
-    };
-    // Confirm it is one of our GC-managed materialization blocks.
-    if !pyre_object::gc_hook::try_gc_owns_object(block as *mut u8) {
-        return;
-    }
-    let idx = match ctx.trace_ctx.box_value(index) {
-        Some(majit_ir::Value::Int(i)) if i >= 0 => i as usize,
-        _ => {
-            ctx.trace_ctx
-                .try_set_opref_concrete(array, majit_ir::Value::Ref(majit_ir::GcRef::NO_CONCRETE));
-            return;
-        }
-    };
-    let cap = unsafe { pyre_object::object_array::items_block_capacity(block) };
-    if idx >= cap {
-        ctx.trace_ctx
-            .try_set_opref_concrete(array, majit_ir::Value::Ref(majit_ir::GcRef::NO_CONCRETE));
-        return;
-    }
-    let elem = match ctx.trace_ctx.box_value(value) {
-        Some(majit_ir::Value::Ref(r)) if r != majit_ir::GcRef::NO_CONCRETE => {
-            r.as_usize() as pyre_object::PyObjectRef
-        }
-        _ => {
-            ctx.trace_ctx
-                .try_set_opref_concrete(array, majit_ir::Value::Ref(majit_ir::GcRef::NO_CONCRETE));
-            return;
-        }
-    };
-    unsafe {
-        let base = pyre_object::object_array::items_block_items_base(block);
-        *base.add(idx) = elem;
-    }
-    // Old→young barrier: the materialization block may have been promoted to
-    // old-gen while `elem` is still young (the construction-barrier gap). A
-    // nursery block carries no TRACK_YOUNG_PTRS so the barrier is a no-op.
-    pyre_object::gc_hook::try_gc_write_barrier(block as *mut u8);
-}
-
-/// `setfield_gc_<i|r>/<rid|rrd>` handler: read box (r-reg), valuebox
-/// (i or r reg per `value_bank`), descr operand, then either skip
-/// the IR emission (cache says the same value is already there) or
-/// record `OpCode::SetfieldGc` and write through to the heapcache.
-///
-/// RPython parity: `pyjitpl.py:973-988 _opimpl_setfield_gc_any`:
-///
-///   upd = heapcache.get_field_updater(box, fielddescr)
-///   if upd.currfieldbox is valuebox:
-///       return                       # cache hit, no IR
-///   self.metainterp.execute_and_record(rop.SETFIELD_GC, fielddescr,
-///                                       box, valuebox)
-///   upd.setfield(valuebox)
-///
-/// **Alias-clearing writeback**: goes through
-/// `HeapCache::setfield_cached` instead of `getfield_now_known`. The
-/// difference is the alias-clearing semantic that RPython's
-/// `FieldUpdater.setfield()` carries (heapcache.py:142-143 routes to
-/// `CacheEntry.do_write_with_aliasing`):
-///
-///   `_clear_cache_on_write(seen_alloc)` (heapcache.py:70-77) wipes
-///   `cache_anything` unconditionally and additionally wipes
-///   `cache_seen_allocation` when the write target itself is not
-///   seen-allocated.  This conservatively kills any cached entry whose
-///   source-box might alias the SETFIELD target.
-///
-/// `getfield_now_known` only inserts the new (obj, field, value) tuple
-/// — it does NOT clear sibling entries.  Using it here meant a
-/// subsequent `getfield_gc(other_obj, same_field)` could return a
-/// stale value cached from before the SETFIELD.  Switching to
-/// `setfield_cached` matches `do_write_with_aliasing` exactly.
-///
-/// `value_bank` selects the valuebox source: `'i'` reads
-/// `registers_i[v]`, `'r'` reads `registers_r[v]`.
-fn setfield_gc_via_heapcache(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    value_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Operand layout `<r><v>d`: 1B r-reg(box) + 1B v(value) + 2B descr-index.
-    // For the `c`-coded short form (`setfield_gc_i/rcd`) the value byte is
-    // an inline signed constant (`signedord`, `blackhole.py:123`) read as a
-    // `ConstInt` box instead of an `i`-register slot; obj and descr keep the
-    // `rid` byte positions.
-    let obj = read_ref_reg(code, op, 0, ctx)?;
-    let valuebox = match value_bank {
-        'i' => read_int_reg(code, op, 1, ctx)?,
-        'r' => read_ref_reg(code, op, 1, ctx)?,
-        'f' => read_float_reg(code, op, 1, ctx)?,
-        'c' => OpRef::ConstInt(code[op.pc + 2] as i8 as i64),
-        _ => unreachable!("value_bank must be 'i', 'r', 'f' or 'c'"),
-    };
-    let descr = read_descr(code, op, 2, ctx)?;
-    let descr_index = descr.index();
-
-    // Cache hit: if the heapcache already records `valuebox` as the
-    // current value of `(obj, descr)`, the SETFIELD_GC is redundant —
-    // skip recording. RPython pyjitpl.py:973-979 _opimpl_setfield_gc_any:
-    //   if upd.currfieldbox is valuebox:
-    //       self.metainterp.staticdata.profiler.count_ops(rop.SETFIELD_GC, Counters.HEAPCACHED_OPS)
-    //       return
-    let is_redundant = ctx
-        .trace_ctx
-        .heapcache_getfield_cached(obj, descr_index)
-        .map(|b| b)
-        == Some(valuebox);
-    if is_redundant {
-        ctx.trace_ctx.profiler().count_ops(
-            OpCode::SetfieldGc,
-            majit_metainterp::counters::HEAPCACHED_OPS,
-        );
-    } else {
-        ctx.trace_ctx
-            .record_op_with_descr(OpCode::SetfieldGc, &[obj, valuebox], descr);
-        // Write-through with alias-clearing semantics
-        // (`heapcache.py:90-94 do_write_with_aliasing`).  Mirrors
-        // `upd.setfield(valuebox)` (heapcache.py:142).
-        ctx.trace_ctx
-            .heapcache_setfield_cached(obj, descr_index, valuebox);
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `getfield_gc_<i|r>/rd>X` handler: read a Ref-bank source register
-/// + descr operand, consult the heapcache, and either return the
-/// cached field box (no IR op recorded) or record the appropriate
-/// `OpCode::GetfieldGc<I|R>` op and update the cache.
-///
-/// RPython parity: `pyjitpl.py:855-882 opimpl_getfield_gc_<i|r>` →
-/// `_opimpl_getfield_gc_any_pureornot` (`pyjitpl.py:929-950`).
-/// RPython has a ConstPtr+is_always_pure() fast path at lines 856-860
-/// that fires `executor.execute(cpu, metainterp, opnum, fielddescr,
-/// box)` and returns `ConstInt/ConstFloat/ConstPtr(resvalue)` —
-/// recording NO trace op (the value is directly substituted as a Const
-/// literal). The walker's `executor.execute` counterpart is
-/// `field_sanity_load`, so the fast path is implemented: a constant
-/// source register through an always-pure descr folds to the loaded
-/// value as a Const literal with no recorded op.
-///
-/// Walker behaviour mirrors `_opimpl_getfield_gc_any_pureornot`
-/// uniformly: heapcache hit returns the cached box (no IR op);
-/// heapcache miss records `GetfieldGc<I|R>` (non-pure variant) +
-/// writes through. The optimizer's always-pure pass later folds the
-/// non-pure read into `GetfieldGcPure*` based on `descr.is_always_pure()`,
-/// which is `OpHelpers.getfield_pure_for_descr` (resoperation.py:
-/// 1284-1289) parity. Walker emitting Pure variants directly would be
-/// a TODO since RPython's opimpl_* never emits the Pure
-/// opcodes; they're an optimizer-rewrite artifact.
-///
-/// `dst_bank` selects the result bank: `'i'` writes `registers_i[dst]`,
-/// `'r'` writes `registers_r[dst]`.
-fn getfield_gc_via_heapcache(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    opcode: OpCode,
-    dst_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Operand layout `rd>X`: 1B r-reg + 2B descr-index + 1B dst.
-    let obj = read_ref_reg(code, op, 0, ctx)?;
-    let descr = read_descr(code, op, 1, ctx)?;
-    let descr_index = descr.index();
-
-    // ConstPtr + always-pure fast path (pyjitpl.py:856-860): a constant
-    // source through an immutable descr loads the field now and
-    // substitutes the value as a Const literal, recording no op.
-    let const_pure_result = if obj.is_constant() && descr.is_always_pure() {
-        let load_type = match opcode {
-            OpCode::GetfieldGcI | OpCode::GetfieldGcPureI => Some(majit_ir::Type::Int),
-            OpCode::GetfieldGcR | OpCode::GetfieldGcPureR => Some(majit_ir::Type::Ref),
-            OpCode::GetfieldGcF | OpCode::GetfieldGcPureF => Some(majit_ir::Type::Float),
-            _ => None,
-        };
-        let struct_ptr = match ctx.trace_ctx.box_value(obj) {
-            Some(majit_ir::Value::Ref(struct_ref)) => {
-                let p = struct_ref.0 as i64;
-                (p != 0 && p != usize::MAX as i64).then_some(p)
-            }
-            _ => None,
-        };
-        match (load_type, struct_ptr) {
-            (Some(ty), Some(p)) => {
-                ctx.trace_ctx
-                    .field_sanity_load(p, &descr, ty)
-                    .map(|v| match v {
-                        majit_ir::Value::Int(n) => ctx.trace_ctx.const_int(n),
-                        majit_ir::Value::Ref(r) => ctx.trace_ctx.const_ref(r.0 as i64),
-                        majit_ir::Value::Float(f) => ctx.trace_ctx.const_float(f.to_bits() as i64),
-                        _ => unreachable!("field_sanity_load returns Int/Ref/Float only"),
-                    })
-            }
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    // heaptracker.py:66 special-cases the `typeptr` field: once a GUARD_CLASS
-    // has pinned an object's class, reading its typeptr yields the known
-    // class constant.  Inside an inline sub-walk the receiver's concrete
-    // pointer often lives only in the register shadow (not the box value), so
-    // the const-pure path above misses; fold the typeptr read straight from
-    // the heapcache's known class instead.  This lets inlined type predicates
-    // (`is_int`/`is_bool`, which read the typeptr and compare it against a
-    // type address) fold during the walk.
-    let is_typeptr_field = descr
-        .as_field_descr()
-        .is_some_and(|fd| fd.offset() == pyre_object::pyobject::OB_TYPE_OFFSET);
-    let typeptr_const = if ctx.fbw_mode.inline_subwalk && !obj.is_constant() && is_typeptr_field {
-        let known = ctx.trace_ctx.heap_cache().get_known_class(obj);
-        match (known, opcode) {
-            (Some(cls), OpCode::GetfieldGcI | OpCode::GetfieldGcPureI) => {
-                Some(ctx.trace_ctx.const_int(cls))
-            }
-            (Some(cls), OpCode::GetfieldGcR | OpCode::GetfieldGcPureR) => {
-                Some(ctx.trace_ctx.const_ref(cls))
-            }
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    let result = if let Some(folded) = typeptr_const {
-        folded
-    } else if let Some(constant) = const_pure_result {
-        constant
-    } else if let Some(cached) = ctx.trace_ctx.heapcache_getfield_cached(obj, descr_index) {
-        // Cache hit (RPython pyjitpl.py:929-947 _opimpl_getfield_gc_any_pureornot):
-        //   if upd.currfieldbox is not None:
-        //       self.metainterp.staticdata.profiler.count_ops(rop.GETFIELD_GC_I, Counters.HEAPCACHED_OPS)
-        //       return upd.currfieldbox
-        // RPython hardcodes `GETFIELD_GC_I` for the count regardless of
-        // the actual rop variant (`_i` / `_r` / `_f`); match the
-        // hardcode for profiling parity.
-        ctx.trace_ctx.profiler().count_ops(
-            OpCode::GetfieldGcI,
-            majit_metainterp::counters::HEAPCACHED_OPS,
-        );
-        cached
-    } else {
-        // Recording a `getfield_gc_*` whose FieldDescr lacks a parent_descr
-        // backreference would later crash the optimizer's
-        // `ensure_ptr_info_arg0` (`optimizer.py:478-484`).  Inside a sub-walk
-        // abort gracefully instead, so the trace falls back to the interpreter
-        // rather than carrying an op the optimizer cannot lower.  The fold
-        // paths above (typeptr / const-pure / cache-hit) record nothing, so
-        // they are unaffected; production sub-walks never reach this (they
-        // would already panic).
-        if ctx.fbw_mode.inline_subwalk
-            && matches!(
-                opcode,
-                OpCode::GetfieldGcI
-                    | OpCode::GetfieldGcR
-                    | OpCode::GetfieldGcF
-                    | OpCode::GetfieldGcPureI
-                    | OpCode::GetfieldGcPureR
-                    | OpCode::GetfieldGcPureF
-            )
-            && descr
-                .as_field_descr()
-                .is_some_and(|fd| fd.get_parent_descr().is_none())
-        {
-            return Err(DispatchError::FieldDescrMissingParentDescr { pc: op.pc });
-        }
-        // Cache miss — record op + write through.  `box_value`
-        // resolves the Box.value chain PyPy reads off
-        // `box.getref_base()` in `executor.do_getfield_gc_*`
-        // (`executor.py:188`); the sanity load fires whenever the
-        // struct pointer is known (Const, vable shadow, or stamped),
-        // mirroring `pyjitpl.py:948-949 resbox = execute_with_descr(...);
-        // upd.getfield_now_known(resbox)`.
-        let resbox = ctx
-            .trace_ctx
-            .record_op_with_descr(opcode, &[obj], descr.clone());
-        let load_type = match opcode {
-            OpCode::GetfieldGcI | OpCode::GetfieldGcPureI => Some(majit_ir::Type::Int),
-            OpCode::GetfieldGcR | OpCode::GetfieldGcPureR => Some(majit_ir::Type::Ref),
-            OpCode::GetfieldGcF | OpCode::GetfieldGcPureF => Some(majit_ir::Type::Float),
-            _ => None,
-        };
-        let live_value = if let (Some(ty), Some(majit_ir::Value::Ref(struct_ref))) =
-            (load_type, ctx.trace_ctx.box_value(obj))
-        {
-            let struct_ptr = struct_ref.0 as i64;
-            if struct_ptr != usize::MAX as i64 && struct_ptr != 0 {
-                ctx.trace_ctx.field_sanity_load(struct_ptr, &descr, ty)
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-        // Stamp the loaded value as the Box.value of the recorded
-        // result so subsequent reads (cache hits + non-Const
-        // `concrete_of_opref` consumers) see the real runtime
-        // concrete instead of the GcRef(usize::MAX) sentinel.
-        if let Some(live_value) = live_value {
-            ctx.trace_ctx.set_opref_concrete(resbox, live_value);
-        }
-        ctx.trace_ctx
-            .heapcache_getfield_now_known(obj, descr_index, resbox);
-        resbox
-    };
-
-    let dst = code[op.pc + 4] as usize;
-    // concrete_of_opref derivation: derive shadow concrete via `concrete_of_opref` so a
-    // constant-folded predecessor (e.g. `binop_int_record` having
-    // stamped this OpRef in OpRef concrete stamping) propagates through.  RPython
-    // `Box.value` parity: `pyjitpl.py:executor.py` per-opcode LLOp
-    // stamps `box.value` post-exec; pyre's `concrete_of_opref` reads
-    // that channel.  Null fallback preserves the prior unknown-result
-    // behaviour for cache-miss recorded ops.
-    let concrete_for_shadow = concrete_from_recorded_opref(ctx, result);
-    match dst_bank {
-        'i' => {
-            write_int_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'r' => {
-            write_ref_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'f' => {
-            let len = ctx.registers_f.len();
-            let slot = ctx
-                .registers_f
-                .get_mut(dst)
-                .ok_or(DispatchError::RegisterOutOfRange {
-                    pc: op.pc,
-                    reg: dst,
-                    len,
-                    bank: "f",
-                })?;
-            *slot = result;
-        }
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `virtualizable_gen.rs:33-60` pyre PyFrame static-field order
-/// `[last_instr, pycode, valuestackdepth, debugdata, lastblock, w_globals]`.
-const VABLE_CODE_FIELD_IDX: usize = 1;
-const VABLE_NAMESPACE_FIELD_IDX: usize = 5;
-
 /// Resolve the top-level walk's live frame code when the promoted `pycode`
 /// green has no concrete shadow.
 ///
@@ -4373,667 +2719,6 @@ fn try_resolve_inline_callee_static_field(
         ConcreteValue::Ref(const_ptr as pyre_object::PyObjectRef),
     )?;
     Ok(Some((DispatchOutcome::Continue, op.next_pc)))
-}
-
-/// `getfield_vable_<i|r|f>/rd>X` handler. Operand layout `rd>X`:
-/// 1B r-reg(vable_box) + 2B descr(field) + 1B X-dst.
-///
-/// RPython parity: `pyjitpl.py:1167-1186 opimpl_getfield_vable_{i,r,f}`:
-///
-///   def opimpl_getfield_vable_i(self, box, fielddescr, pc):
-///       if self._nonstandard_virtualizable(pc, box, fielddescr):
-///           return self.opimpl_getfield_gc_i(box, fielddescr)
-///       self.metainterp.check_synchronized_virtualizable()
-///       index = self._get_virtualizable_field_index(fielddescr)
-///       return self.metainterp.virtualizable_boxes[index]
-///
-/// The walker delegates to the orthodox `TraceCtx::vable_getfield_{int,
-/// ref,float}` ports (`majit-metainterp/src/trace_ctx.rs:1715, 1801,
-/// 1839`) which already implement the full
-/// `_nonstandard_virtualizable` check + heapcache-aware GETFIELD_GC
-/// fallback + `virtualizable_boxes[index]` cache read.  Only the OpRef
-/// component of the `(OpRef, Value)` tuple is meaningful here, since
-/// register banks carry only OpRefs — the concrete `Value` is tracked
-/// via the per-step concrete frame snapshot.  `dst_bank` selects the result bank
-/// (`'i'`/`'r'`/`'f'`) the walker writes back into, mirroring
-/// `getfield_gc_via_heapcache`'s shape.
-fn getfield_vable_via_metainterp(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    dst_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    let obj = read_ref_reg(code, op, 0, ctx)?;
-    // #68/gap-10: inside an inlined-callee sub-walk, a scalar getfield_vable_r
-    // of the callee's namespace(idx5)/pycode(idx1) must resolve to the callee's
-    // compile-time `InlineCalleeConsts` whether the callee frame is unseeded
-    // (strict path, handled by the `obj.is_none()` branch below) OR a seeded
-    // virtual frame (multiframe path). On the multiframe path the read hits the
-    // seeded frame box but the heapcache forward MISSES — the codewriter's
-    // per-fn vable descr identity differs from the seeding descr — so it would
-    // record a non-const `GetfieldGcR`, leaving the LOAD_GLOBAL fold's namespace
-    // operand non-concrete (the residual that the nested-unjournaled abort
-    // declines, e.g. nbody `advance()`'s `len`). The consts route is
-    // descr-identity-independent (`const_ref`) and only intercepts the two
-    // static Ref fields (`try_resolve_inline_callee_static_field` returns `None`
-    // for everything else), so it is safe to consult on the seeded path too.
-    if fbw_inline_nsfold_enabled() {
-        if let Some(resolved) = try_resolve_inline_callee_static_field(code, op, ctx, dst_bank)? {
-            return Ok(resolved);
-        }
-    }
-    // RPython's `box` is always a live virtualizable-frame box. An
-    // unseeded walker Ref register holds `OpRef::None` (`raw() ==
-    // u32::MAX`); feeding it into the metainterp vable path would resize
-    // the heapcache flag vector to 16 GiB. Bail to a trace abort instead.
-    if obj.is_none() {
-        // Path-1 (#68): an inlined callee reading a scalar field off its
-        // OWN unseeded portal frame.  A resolvable static field (`w_globals`
-        // namespace for a LOAD_GLOBAL, `pycode` promote-to-const) folds to
-        // the callee constant; anything else is declined up-front by
-        // `callee_fast_path_inlinable`, so an unresolved field here is a
-        // genuine unseeded-box error.
-        if let Some(resolved) = try_resolve_inline_callee_static_field(code, op, ctx, dst_bank)? {
-            return Ok(resolved);
-        }
-        return Err(DispatchError::VableBoxNotSeeded { pc: op.pc });
-    }
-    let descr = read_descr(code, op, 1, ctx)?;
-
-    // R7 parity: RPython `opimpl_getfield_vable_{i,r,f}(box, fielddescr,
-    // pc)` threads orgpc through `_nonstandard_virtualizable(pc, ...)`
-    // (pyjitpl.py:1167-1186 + :1137).  Pyre's walker has the matching
-    // JitCode PC in `op.pc`; pass it through so the helper signature
-    // stays line-by-line equivalent even if `is_nonstandard_virtualizable`
-    // currently ignores the pc at the leaf (`trace_ctx.rs let _ = pc;`).
-    let pc = op.pc;
-    // Concrete struct pointer for pyjitpl.py:934-945 cache-hit sanity
-    // check.  The walker keeps a parallel concrete Ref-bank shadow;
-    // thread the same live pointer that RPython's `box.getref_base()`
-    // would expose to `executor.execute(...)`.
-    let vable_struct_ptr = match read_ref_reg_concrete(code, op, 0, ctx) {
-        ConcreteValue::Ref(ptr) => ptr as i64,
-        ConcreteValue::Null => 0,
-        ConcreteValue::Int(_) | ConcreteValue::Float(_) | ConcreteValue::Bool(_) => 0,
-    };
-    let guards_before = ctx.trace_ctx.num_guards();
-    let (result, shadow_value) = match dst_bank {
-        'i' => ctx
-            .trace_ctx
-            .vable_getfield_int(pc, obj, vable_struct_ptr, descr),
-        'r' => ctx
-            .trace_ctx
-            .vable_getfield_ref(pc, obj, vable_struct_ptr, descr),
-        'f' => ctx
-            .trace_ctx
-            .vable_getfield_float(pc, obj, vable_struct_ptr, descr),
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    };
-    walker_capture_inline_nonstandard_vable_guard(ctx, op.pc, guards_before)?;
-    // RPython `opimpl_getfield_vable_{i,r,f}` returns
-    // `virtualizable_boxes[index]` (`pyjitpl.py:1186`) — a Box whose
-    // `_resint`/`_resref`/`_resfloat` is filled at construction time.
-    // `box.getint()` returns the live value without any side-lookup.
-    // Pyre splits OpRef↔concrete into a side table; mirror the Box.value
-    // contract by stamping the read result's concrete into
-    // `opref_concrete` so `concrete_of_opref(result)` honors the same
-    // contract for downstream consumers (`goto_if_not/iL`,
-    // `switch/id`, `int_*` arithmetic).  The non-standard heapcache
-    // path inside `vable_getfield_int` already does the same stamp;
-    // the standard path returns the cached `(opref, value)` pair
-    // without stamping.  `None` means no live concrete is available
-    // for this slot — skip to match the heapcache path's gating.
-    if let Some(shadow_value) = shadow_value {
-        ctx.trace_ctx.set_opref_concrete(result, shadow_value);
-    }
-    let dst = code[op.pc + 4] as usize;
-    // concrete_of_opref derivation: derive shadow concrete via `concrete_of_opref`.  The
-    // `vable_getfield_*` helpers in `TraceCtx` already populate the
-    // concrete shadow for virtualizable-resident fields via the
-    // `standard_virtualizable_box()`/`virtualizable_boxes` channel,
-    // and feed `set_opref_concrete` on the GETFIELD_GC fallback for
-    // non-vable structs — both surface through this lookup.
-    let concrete_for_shadow = concrete_from_recorded_opref(ctx, result);
-    match dst_bank {
-        'i' => {
-            write_int_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'r' => {
-            write_vable_field_ref_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-        }
-        'f' => {
-            let len = ctx.registers_f.len();
-            let slot = ctx
-                .registers_f
-                .get_mut(dst)
-                .ok_or(DispatchError::RegisterOutOfRange {
-                    pc: op.pc,
-                    reg: dst,
-                    len,
-                    bank: "f",
-                })?;
-            *slot = result;
-        }
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `setfield_vable_<i|r|f>/r<v>d` handler. Operand layout:
-/// 1B r-reg(vable_box) + 1B <v>-reg(value) + 2B descr(field).
-/// No dst byte (set, not get).
-///
-/// RPython parity: `pyjitpl.py:1188-1199 _opimpl_setfield_vable`:
-///
-///   def _opimpl_setfield_vable(self, box, valuebox, fielddescr, pc):
-///       if self._nonstandard_virtualizable(pc, box, fielddescr):
-///           return self._opimpl_setfield_gc_any(box, valuebox, fielddescr)
-///       index = self._get_virtualizable_field_index(fielddescr)
-///       self.metainterp.virtualizable_boxes[index] = valuebox
-///       self.metainterp.synchronize_virtualizable()
-///       # XXX only the index'th field needs to be synchronized, really
-///
-/// The walker delegates to `TraceCtx::vable_setfield`
-/// (`majit-metainterp/src/trace_ctx.rs:1759`) which implements the
-/// full `_nonstandard_virtualizable` -> SETFIELD_GC fallback +
-/// `virtualizable_boxes[index] = valuebox` write + `synchronize_virtualizable`
-/// mirror.  The concrete `Value` is reconstructed via
-/// `TraceCtx::concrete_of_opref` (matches the trait-leg's
-/// `pyjitpl/dispatch.rs:1608-1609` shape `let (value, concrete) =
-/// self.read_<bank>_reg(src); ctx.vable_setfield(...)`).
-///
-/// `value_bank` selects the value register bank (`'i'`/`'r'`/`'f'`),
-/// mirroring `setfield_gc_via_heapcache`'s parameter shape.
-fn setfield_vable_via_metainterp(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    value_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Strict fresh-frame fold: a scalar setfield to the current inline level's
-    // own (unseeded) portal frame is a virtual-field write — `valuestackdepth`
-    // / `last_instr` sync on a branchless leaf that resumes at the caller
-    // boundary — so it folds to a no-op, emitting no SETFIELD_GC.  The
-    // `fresh_virtualizable` OptVirtualize elision (jtransform.py:990-993).
-    let fold_frame_reg = fbw_strict_fold_frame_reg(ctx);
-    if fold_frame_reg != u16::MAX && code[op.pc + 1] as u16 == fold_frame_reg {
-        return Ok((DispatchOutcome::Continue, op.next_pc));
-    }
-    let obj = read_ref_reg(code, op, 0, ctx)?;
-    // Same unseeded-register guard as `getfield_vable_via_metainterp`:
-    // a `None` box would resize the heapcache flag vector to 16 GiB.
-    if obj.is_none() {
-        return Err(DispatchError::VableBoxNotSeeded { pc: op.pc });
-    }
-    let value = match value_bank {
-        'i' => read_int_reg(code, op, 1, ctx)?,
-        'r' => read_ref_reg(code, op, 1, ctx)?,
-        'f' => read_float_reg(code, op, 1, ctx)?,
-        _ => unreachable!("value_bank must be 'i', 'r' or 'f'"),
-    };
-    let descr = read_descr(code, op, 2, ctx)?;
-    let concrete = ctx.trace_ctx.concrete_of_opref(value);
-    // R7 parity: pyjitpl.py:1188-1199 `_opimpl_setfield_vable(box,
-    // valuebox, fielddescr, pc)` threads orgpc through
-    // `_nonstandard_virtualizable(pc, ...)`; walker has `op.pc` for the
-    // JitCode PC, pass through.
-    let guards_before = ctx.trace_ctx.num_guards();
-    ctx.trace_ctx
-        .vable_setfield(op.pc, obj, descr, value, concrete);
-    walker_capture_inline_nonstandard_vable_guard(ctx, op.pc, guards_before)?;
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// Resolve the `(fdescr, adescr)` virtualizable-array descr pair from a
-/// `(VableArray, Array)` jitcode descr-pool pair.
-///
-/// RPython parity: `MIFrame.vable_array_index_pair_at`
-/// (`blackhole.rs:1613-1628`) reads the two 2-byte descr-pool indices,
-/// asserts they resolve to `(BhDescr::VableArray { index }, BhDescr::Array
-/// { .. })`, and yields the `index`. `vable_array_descrs`
-/// (`pyjitpl/dispatch.rs:845-856`) then maps that index through the
-/// jitdriver's `VirtualizableInfo` to the canonical `array_field_descrs`
-/// /`array_descrs` pair.  The walker must use the *vinfo* descrs (not the
-/// raw jitcode-pool descrs) because `TraceCtx::vable_array_flat_index`
-/// keys on `array_field_by_descr` (descr identity, `virtualizable.rs:602`).
-///
-/// `field_offset` / `array_offset` are byte offsets (from `op.pc + 1`) of
-/// the two descr operands, matching the per-op argcode layout.
-fn vable_array_descrs_from_jitcode(
-    code: &[u8],
-    op: &DecodedOp,
-    field_offset: usize,
-    array_offset: usize,
-    ctx: &WalkContext<'_, '_>,
-) -> Result<(DescrRef, DescrRef), DispatchError> {
-    let read_pool_idx = |off: usize| {
-        let lo = code[op.pc + 1 + off] as usize;
-        let hi = code[op.pc + 1 + off + 1] as usize;
-        lo | (hi << 8)
-    };
-    let field_idx = read_pool_idx(field_offset);
-    let array_pool_idx = read_pool_idx(array_offset);
-    // RPython `MIFrame.vable_array_index_pair_at` reads `self.descrs[idx]`
-    // — pyre's single per-walk pool, selected by `ctx.raw_descrs`
-    // (global `ALL_DESCRS` for arm walks, per-`CodeObject` `exec.descrs`
-    // for full-body walks).
-    let array_field_index = match (
-        ctx.raw_descrs.bh_descr_at(field_idx),
-        ctx.raw_descrs.bh_descr_at(array_pool_idx),
-    ) {
-        (
-            Some(majit_translate::jitcode::BhDescr::VableArray { index }),
-            Some(majit_translate::jitcode::BhDescr::Array { .. }),
-        ) => *index,
-        _ => {
-            if std::env::var("PYRE_FBW_INLINE_DIAG").is_ok() {
-                eprintln!(
-                    "[vable-arr-malformed] pc={} pool_len={:?} field_idx={field_idx} \
-                     field={:?} array_idx={array_pool_idx} array={:?}",
-                    op.pc,
-                    ctx.raw_descrs.len(),
-                    ctx.raw_descrs.bh_descr_at(field_idx),
-                    ctx.raw_descrs.bh_descr_at(array_pool_idx),
-                );
-            }
-            return Err(DispatchError::VableArrayDescrMalformed {
-                pc: op.pc,
-                field_idx,
-                array_idx: array_pool_idx,
-            });
-        }
-    };
-    let info = ctx
-        .trace_ctx
-        .virtualizable_info()
-        .ok_or(DispatchError::VableArrayMissingVirtualizableInfo { pc: op.pc })?;
-    let fdescr = info
-        .array_field_descrs()
-        .get(array_field_index)
-        .cloned()
-        .ok_or(DispatchError::VableArrayIndexOutOfRange {
-            pc: op.pc,
-            index: array_field_index,
-        })?;
-    let adescr = info.array_descrs.get(array_field_index).cloned().ok_or(
-        DispatchError::VableArrayIndexOutOfRange {
-            pc: op.pc,
-            index: array_field_index,
-        },
-    )?;
-    Ok((fdescr, adescr))
-}
-
-/// `getarrayitem_vable_<i|r|f>/ridd>X` handler. Operand layout `ridd>X`:
-/// 1B r-reg(vable) + 1B i-reg(index) + 2B fdescr(VableArray) + 2B
-/// adescr(Array) + 1B X-dst.
-///
-/// RPython parity: `pyjitpl.py:1218-1234 _opimpl_getarrayitem_vable`
-/// (`opimpl_getarrayitem_vable_{i,r,f}`).  Delegates to
-/// `TraceCtx::vable_getarrayitem_{int,ref,float}_indexed`
-/// (`trace_ctx.rs:2982/3043/3099`) which implements the
-/// `_nonstandard_virtualizable` GETFIELD_GC + GETARRAYITEM_GC fallback
-/// and the standard-vable `virtualizable_boxes[index]` cache read.
-/// Mirrors `getfield_vable_via_metainterp`'s concrete-stamp + dst-write
-/// shape; the trait counterpart is `pyjitpl/dispatch.rs:1909-1977`.
-fn getarrayitem_vable_via_metainterp(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    dst_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Strict fresh-frame fold: when this op reads the current inline level's
-    // own (unseeded) portal frame, resolve it register-to-register through the
-    // per-slot OpRef shadow and emit NO GC op — the `fresh_virtualizable` case
-    // (jtransform.py:990-993 `is_virtualizable_getset` returns `False`).  The
-    // param-store seed dominates every read of a branchless leaf's own frame,
-    // so the slot is present; a genuinely-unbound read falls through to the
-    // `VableBoxNotSeeded` abort below (never the 16 GiB metainterp path).
-    let fold_frame_reg = fbw_strict_fold_frame_reg(ctx);
-    if fold_frame_reg != u16::MAX && code[op.pc + 1] as u16 == fold_frame_reg {
-        let index = read_int_reg(code, op, 1, ctx)?;
-        if let Some(majit_ir::Value::Int(slot)) = ctx.trace_ctx.concrete_of_opref(index) {
-            if let Some(result) = ctx
-                .callee_shadow
-                .as_ref()
-                .and_then(|shadow| shadow.opref.get(&slot).copied())
-            {
-                let dst = code[op.pc + 7] as usize;
-                let concrete = concrete_from_recorded_opref(ctx, result);
-                match dst_bank {
-                    'i' => write_int_reg(ctx, op.pc, dst, result, concrete)?,
-                    'r' => write_ref_reg(ctx, op.pc, dst, result, concrete)?,
-                    'f' => {
-                        let len = ctx.registers_f.len();
-                        let slot_ref = ctx.registers_f.get_mut(dst).ok_or(
-                            DispatchError::RegisterOutOfRange {
-                                pc: op.pc,
-                                reg: dst,
-                                len,
-                                bank: "f",
-                            },
-                        )?;
-                        *slot_ref = result;
-                    }
-                    _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-                }
-                return Ok((DispatchOutcome::Continue, op.next_pc));
-            }
-        }
-    }
-    let vable = read_ref_reg(code, op, 0, ctx)?;
-    // An unseeded walker Ref register holds `OpRef::None` (`raw() ==
-    // u32::MAX`); feeding it into the metainterp vable path would resize
-    // the heapcache flag vector to 16 GiB. Bail to a trace abort, mirroring
-    // the scalar `getfield_vable_via_metainterp` guard.
-    if vable.is_none() {
-        return Err(DispatchError::VableBoxNotSeeded { pc: op.pc });
-    }
-    let index = read_int_reg(code, op, 1, ctx)?;
-    // pyjitpl.py:1206 `indexbox.getint()` — the array slot is chosen from
-    // the concrete index. Fail loud if the walker can't resolve it.
-    let index_value = match ctx.trace_ctx.concrete_of_opref(index) {
-        Some(Value::Int(v)) => v,
-        _ => {
-            return Err(DispatchError::VableArrayIndexNotConcrete {
-                pc: op.pc,
-                value: index,
-            });
-        }
-    };
-    let (fdescr, adescr) = vable_array_descrs_from_jitcode(code, op, 2, 4, ctx)?;
-    let guards_before = ctx.trace_ctx.num_guards();
-    let (result, shadow_value) = match dst_bank {
-        'i' => ctx.trace_ctx.vable_getarrayitem_int_indexed(
-            op.pc,
-            vable,
-            index,
-            index_value,
-            fdescr,
-            adescr,
-        ),
-        'r' => ctx.trace_ctx.vable_getarrayitem_ref_indexed(
-            op.pc,
-            vable,
-            index,
-            index_value,
-            fdescr,
-            adescr,
-        ),
-        'f' => ctx.trace_ctx.vable_getarrayitem_float_indexed(
-            op.pc,
-            vable,
-            index,
-            index_value,
-            fdescr,
-            adescr,
-        ),
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    };
-    walker_capture_inline_nonstandard_vable_guard(ctx, op.pc, guards_before)?;
-    let shadow_value = shadow_value.unwrap_or(Value::Void);
-    // When the read missed every concrete channel (`Void`) but we are inside
-    // an inlined callee, fall back to the per-frame concrete-locals shadow —
-    // a loop-carried local read after a may-force op in the loop body has no
-    // heapcache entry this pass, yet its recording-time concrete is known.
-    let shadow_value = if matches!(shadow_value, Value::Void) {
-        ctx.callee_shadow
-            .as_ref()
-            .and_then(|shadow| shadow.concrete.get(&index_value).copied())
-            .filter(|entry| entry.frame_reg == code[op.pc + 1] as u16)
-            .map(|entry| entry.value)
-            .unwrap_or(Value::Void)
-    } else {
-        shadow_value
-    };
-    // Mirror `getfield_vable_via_metainterp`: stamp the read result's
-    // concrete so `concrete_of_opref(result)` honors the Box.value
-    // contract for downstream consumers; `Value::Void` = no live concrete.
-    //
-    // An indexed read hands back the STORED box, so the stamp is a write onto a
-    // box that may already carry a live concrete. A frame whose array slot is
-    // still an unmaterialized hole reads NULL while the real Ref lives in the
-    // guard's register file, so stamping that NULL would clobber the live value
-    // (last write wins) and fold a downstream may-force residual's Ref arg to
-    // NULL -> `MayForceNullRefArgUnsupported`. Skip ONLY that clobber: a NULL
-    // read onto a box with no live Ref still stamps, keeping NULL a real
-    // concrete for the slots that genuinely hold one.
-    let clobbers_live_ref = matches!(shadow_value, Value::Ref(majit_ir::GcRef(0)))
-        && matches!(
-            ctx.trace_ctx.box_value(result),
-            Some(Value::Ref(majit_ir::GcRef(addr))) if addr != 0
-        );
-    if !matches!(shadow_value, Value::Void) && !clobbers_live_ref {
-        ctx.trace_ctx.set_opref_concrete(result, shadow_value);
-    }
-    let dst = code[op.pc + 7] as usize;
-    let concrete_for_shadow = concrete_from_recorded_opref(ctx, result);
-    match dst_bank {
-        'i' => write_int_reg(ctx, op.pc, dst, result, concrete_for_shadow)?,
-        'r' => write_ref_reg(ctx, op.pc, dst, result, concrete_for_shadow)?,
-        'f' => {
-            let len = ctx.registers_f.len();
-            let slot = ctx
-                .registers_f
-                .get_mut(dst)
-                .ok_or(DispatchError::RegisterOutOfRange {
-                    pc: op.pc,
-                    reg: dst,
-                    len,
-                    bank: "f",
-                })?;
-            *slot = result;
-        }
-        _ => unreachable!("dst_bank must be 'i', 'r' or 'f'"),
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `setarrayitem_vable_<i|r|f>/riXdd` handler. Operand layout `riXdd`:
-/// 1B r-reg(vable) + 1B i-reg(index) + 1B X-reg(value) + 2B
-/// fdescr(VableArray) + 2B adescr(Array). No dst byte.
-///
-/// RPython parity: `pyjitpl.py:1236-1247 _opimpl_setarrayitem_vable`.
-/// Delegates to `TraceCtx::vable_setarrayitem_indexed`
-/// (`trace_ctx.rs:3153`) which implements the `_nonstandard_virtualizable`
-/// SETARRAYITEM_GC fallback + the standard-vable
-/// `virtualizable_boxes[index] = valuebox` + `synchronize_virtualizable`.
-/// Trait counterpart: `pyjitpl/dispatch.rs:1978-2052`.
-fn setarrayitem_vable_via_metainterp(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-    value_bank: char,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    // Strict fresh-frame fold (twin of `getarrayitem_vable_via_metainterp`): a
-    // write to the current inline level's own (unseeded) portal frame updates
-    // the per-slot OpRef + concrete shadow and emits NO SETARRAYITEM_GC.  Both
-    // local stores (`index < nlocals`) and operand-stack pushes (`index >=
-    // nlocals`) on the fresh frame are pure mirror writes here — the value
-    // lives in an SSA register; the vable array is folded away.
-    let fold_frame_reg = fbw_strict_fold_frame_reg(ctx);
-    if fold_frame_reg != u16::MAX && code[op.pc + 1] as u16 == fold_frame_reg {
-        let index = read_int_reg(code, op, 1, ctx)?;
-        if let Some(majit_ir::Value::Int(slot)) = ctx.trace_ctx.concrete_of_opref(index) {
-            let value = match value_bank {
-                'i' => read_int_reg(code, op, 2, ctx)?,
-                'r' => read_ref_reg(code, op, 2, ctx)?,
-                'f' => read_float_reg(code, op, 2, ctx)?,
-                _ => unreachable!("value_bank must be 'i', 'r' or 'f'"),
-            };
-            let concrete = ctx
-                .trace_ctx
-                .concrete_of_opref(value)
-                .unwrap_or(majit_ir::Value::Void);
-            if let Some(shadow) = ctx.callee_shadow.as_mut() {
-                shadow.set_opref(slot, value);
-                shadow.set_concrete(fold_frame_reg, slot, concrete);
-            }
-            return Ok((DispatchOutcome::Continue, op.next_pc));
-        }
-    }
-    let vable = read_ref_reg(code, op, 0, ctx)?;
-    // See `getarrayitem_vable_via_metainterp`: an unseeded `OpRef::None`
-    // vable would resize the heapcache flag vector to 16 GiB; bail instead.
-    if vable.is_none() {
-        return Err(DispatchError::VableBoxNotSeeded { pc: op.pc });
-    }
-    let index = read_int_reg(code, op, 1, ctx)?;
-    let index_value = match ctx.trace_ctx.concrete_of_opref(index) {
-        Some(Value::Int(v)) => v,
-        _ => {
-            return Err(DispatchError::VableArrayIndexNotConcrete {
-                pc: op.pc,
-                value: index,
-            });
-        }
-    };
-    let mut value = match value_bank {
-        'i' => read_int_reg(code, op, 2, ctx)?,
-        'r' => read_ref_reg(code, op, 2, ctx)?,
-        'f' => read_float_reg(code, op, 2, ctx)?,
-        _ => unreachable!("value_bank must be 'i', 'r' or 'f'"),
-    };
-    // A STORE_FAST (`index < nlocals`) pops the operand-stack TOS and writes
-    // it into a local slot.  When the popped value lives in an operand-stack
-    // temp that spans a nested loop, the loop-header merge-point seeds no
-    // operand-stack color (`trace.rs` loop-header entry), so the codewriter's
-    // value register reads back `OpRef::NONE` even though the box is live.
-    // The walk-level operand-stack mirror (`vstack_boxes`, the analog of
-    // `MIFrame.registers_r`) is the authoritative kept-stack source and was
-    // maintained across the nested loop; recover the TOS box from it.  Writing
-    // the unbound `NONE` into the vable array otherwise leaves an untyped slot
-    // that fails the guard-snapshot buildability precondition
-    // (`GuardSnapshotVableUntyped`).  Gate on a live mirror + a Ref store into
-    // the local region so a genuine null / non-mirrored slot keeps the legacy
-    // read.
-    if value.is_none() && value_bank == 'r' && ctx.vstack_valid {
-        let full_body_sym = ctx.fbw_mode.snapshot_sym;
-        if !full_body_sym.is_null() {
-            // SAFETY: pointer live for the full-body walk; read-only.
-            let nlocals = unsafe { (*full_body_sym).nlocals as i64 };
-            if index_value >= 0 && index_value < nlocals {
-                if let Some(&tos) = ctx.vstack_boxes.last() {
-                    if !tos.is_none() {
-                        value = tos;
-                    }
-                }
-            }
-        }
-    }
-    let (fdescr, adescr) = vable_array_descrs_from_jitcode(code, op, 3, 5, ctx)?;
-    let concrete = ctx
-        .trace_ctx
-        .concrete_of_opref(value)
-        .unwrap_or(Value::Void);
-    let guards_before = ctx.trace_ctx.num_guards();
-    ctx.trace_ctx.vable_setarrayitem_indexed(
-        op.pc,
-        vable,
-        index,
-        index_value,
-        fdescr,
-        adescr,
-        value,
-        concrete,
-    );
-    // Keep the inline concrete-locals shadow current so a later read of this
-    // slot (after a may-force op clears the heapcache) recovers the concrete.
-    if let Some(shadow) = ctx.callee_shadow.as_mut() {
-        shadow.set_concrete(code[op.pc + 1] as u16, index_value, concrete);
-    }
-    walker_capture_inline_nonstandard_vable_guard(ctx, op.pc, guards_before)?;
-    // A Ref stored to the operand-stack region of the vable array is an
-    // operand-stack push (`pyframe.pushvalue` lowers to
-    // `setarrayitem_vable_r(locals_cells_stack_w, depth, w_obj)`). Retain the
-    // last-write TOS candidate for ordinary single-result opcodes. Method-form
-    // LOAD_ATTR is the one two-result shape that also needs its exact lower
-    // slot mirrored: it writes `[method, self]`, and retaining only `self`
-    // made boundary reconciliation refill the method hole from the stale
-    // pre-LOAD_ATTR receiver. A later guard then resumed CALL with that
-    // receiver as its callee. Scope the positional write to that opcode; the
-    // general boundary model remains authoritative for other push/pop shapes.
-    // Local/cell stores remain excluded by the `index >= nlocals` gate.
-    if value_bank == 'r' && ctx.vstack_valid {
-        let full_body_sym = ctx.fbw_mode.snapshot_sym;
-        if !full_body_sym.is_null() {
-            // SAFETY: pointer live for the full-body walk; read-only.
-            let nlocals = unsafe { (*full_body_sym).nlocals } as i64;
-            if index_value >= nlocals {
-                let method_load = unsafe {
-                    let jitcode = (*full_body_sym).jitcode;
-                    if jitcode.is_null() {
-                        false
-                    } else {
-                        let jitcode = &*jitcode;
-                        if jitcode.payload.code_ptr.is_null() {
-                            false
-                        } else {
-                            pyre_interpreter::decode_instruction_at(
-                                &*jitcode.payload.code_ptr,
-                                ctx.vstack_cur_pypc as usize,
-                            )
-                            .is_some_and(|(instr, op_arg)| match instr
-                            {
-                                pyre_interpreter::Instruction::LoadAttr { namei } => {
-                                    namei.get(op_arg).is_method()
-                                }
-                                _ => false,
-                            })
-                        }
-                    }
-                };
-                if method_load {
-                    let stack_slot = (index_value - nlocals) as usize;
-                    if ctx.vstack_boxes.len() <= stack_slot {
-                        ctx.vstack_boxes.resize(stack_slot + 1, OpRef::NONE);
-                    }
-                    ctx.vstack_boxes[stack_slot] = value;
-                }
-                ctx.vstack_last_ref = value;
-            }
-        }
-    }
-    Ok((DispatchOutcome::Continue, op.next_pc))
-}
-
-/// `arraylen_vable/rdd>i` handler. Operand layout `rdd>i`: 1B r-reg(vable)
-/// + 2B fdescr(VableArray) + 2B adescr(Array) + 1B i-dst.
-///
-/// RPython parity: `pyjitpl.py:1253-1263 opimpl_arraylen_vable`.
-/// Delegates to `TraceCtx::vable_arraylen_vable` (`trace_ctx.rs:3195`)
-/// which implements the `_nonstandard_virtualizable` GETFIELD_GC +
-/// ARRAYLEN_GC fallback and the standard-vable
-/// `ConstInt(get_array_length(...))` read.  Trait counterpart:
-/// `pyjitpl/dispatch.rs:2053-2068`.
-fn arraylen_vable_via_metainterp(
-    code: &[u8],
-    op: &DecodedOp,
-    ctx: &mut WalkContext<'_, '_>,
-) -> Result<(DispatchOutcome, usize), DispatchError> {
-    let vable = read_ref_reg(code, op, 0, ctx)?;
-    // See `getarrayitem_vable_via_metainterp`: an unseeded `OpRef::None`
-    // vable would resize the heapcache flag vector to 16 GiB; bail instead.
-    if vable.is_none() {
-        return Err(DispatchError::VableBoxNotSeeded { pc: op.pc });
-    }
-    let vable_struct_ptr = match read_ref_reg_concrete(code, op, 0, ctx) {
-        ConcreteValue::Ref(ptr) => ptr as i64,
-        ConcreteValue::Null
-        | ConcreteValue::Int(_)
-        | ConcreteValue::Float(_)
-        | ConcreteValue::Bool(_) => 0,
-    };
-    let (fdescr, adescr) = vable_array_descrs_from_jitcode(code, op, 1, 3, ctx)?;
-    let guards_before = ctx.trace_ctx.num_guards();
-    let result = ctx
-        .trace_ctx
-        .vable_arraylen_vable(op.pc, vable, vable_struct_ptr, fdescr, adescr);
-    walker_capture_inline_nonstandard_vable_guard(ctx, op.pc, guards_before)?;
-    let dst = code[op.pc + 6] as usize;
-    let concrete_for_shadow = concrete_from_recorded_opref(ctx, result);
-    write_int_reg(ctx, op.pc, dst, result, concrete_for_shadow)?;
-    Ok((DispatchOutcome::Continue, op.next_pc))
 }
 
 /// Line-by-line port of `pyjitpl.py:1960-1993 MetaInterp._build_allboxes`.
